### 2021-03-23 f87947b
Updated code after running lintr::lint_package()
