### 2021-03-23 689a0ea
Minor fix after lintr::lint_package()
