### 2021-03-23 d576aac
minor fix after lintr::lint_package()
