### 2021-03-23 d0f0b3d
minor change after lintr::lint_package()
changed to snake_case
