### 2021-03-23 a142ce2
Minor changes after lintr::lint_package()
