+++ R/new_equations.R
@@ -27,20 +27,20 @@
 #' @param new_allometry a character string with the allometric equation
 #' @param new_coords a vector or matrix of coordinates (longitude, latitude) of
 #'   the calibration data
-#' @param new_minDBH numerical value, minimum DBH for which the equation is
+#' @param new_min_dbh numerical value, minimum DBH for which the equation is
 #'   valid (in cm). Default is NULL (nothing is added).
-#' @param new_maxDBH numerical value, maximum DBH for which the equation is
+#' @param new_max_dbh numerical value, maximum DBH for which the equation is
 #'   valid (in cm). Default is NULL (nothing is added).
-#' @param new_sampleSize number of measurements with which the allometry was
+#' @param new_sample_size number of measurements with which the allometry was
 #'   calibrated. Default is NULL (nothing is added).
-#' @param new_unitDBH character string with unit of DBH in the equation (either
+#' @param new_unit_dbh character string with unit of DBH in the equation (either
 #'   `cm`, `mm` or `inch`). Default is `cm`.
-#' @param new_unitOutput character string with unit of equation output (either
+#' @param new_unit_output character string with unit of equation output (either
 #'   `g`, `kg`, `Mg` or `lbs` if the output is a mass, or `m` if the output is a
 #'   height).
-#' @param new_inputVar independent variable(s) needed in the allometry. Default
+#' @param new_input_var independent variable(s) needed in the allometry. Default
 #'   is `DBH`, other option is `DBH, H`.
-#' @param new_outputVar dependent variable estimated by the allometry. Default
+#' @param new_output_var dependent variable estimated by the allometry. Default
 #'   is `Total aboveground biomass`.
 #' @param use_height_allom a logical value. In allodb we use Bohn et al. (2014)
 #'   for European sites. User need to provide height allometry when needed to
@@ -55,9 +55,9 @@
 #'   new_taxa = "Faga",
 #'   new_allometry = "exp(-2+log(dbh)*2.5)",
 #'   new_coords = c(-0.07, 46.11),
-#'   new_minDBH = 5,
-#'   new_maxDBH = 50,
-#'   new_sampleSize = 50
+#'   new_min_dbh = 5,
+#'   new_max_dbh = 50,
+#'   new_sample_size = 50
 #' )
 new_equations <- function(subset_taxa = "all",
                           subset_climate = "all",
@@ -77,7 +77,9 @@ new_equations <- function(subset_taxa = "all",
                           new_output_var = "Total aboveground biomass",
                           use_height_allom = TRUE) {
   ## open the equation table and get it in the right format ####
-  new_equations <- allodb::equations
+  equations_orig <- allodb::equations
+  colnames(equations_orig) <- tolower(colnames(equations_orig))
+  new_equations <- equations_orig
 
   suppressWarnings(new_equations$dbh_min_cm <-
                      as.numeric(new_equations$dbh_min_cm))
@@ -165,17 +167,17 @@ new_equations <- function(subset_taxa = "all",
     ## check consistency of inputs
     if (is.null(new_taxa) |
         is.null(new_coords) |
-        is.null(new_minDBH) |
-        is.null(new_maxDBH) |
-        is.null(new_sampleSize))
+        is.null(new_min_dbh) |
+        is.null(new_max_dbh) |
+        is.null(new_sample_size))
       stop(
         "You must provide the taxa, coordinates, DBH range
[…]
