### 2021-03-24 de41e45
add test for last lines of illustrate_allodb code
