+++ R/new_equations.R
@@ -177,7 +177,8 @@ new_equations <- function(subset_taxa = "all",
 
     if (!is.numeric(new_min_dbh) |
         !is.numeric(new_max_dbh) | !is.numeric(new_sample_size))  {
-      stop("new_min_dbh, new_max_dbh, new_sample_size should be numeric values.")
+      stop("new_min_dbh, new_max_dbh, new_sample_size should
+           be numeric values.")
     }
 
     if (is.matrix(new_coords)) {
