### 2021-03-26 45b5d09
Minor change after lintr::lint_package()
