### 2021-03-26 accf996
Revert "Moved opening curly braces"
This reverts commit bb8fe2721131f47cea69477c0d8488ceaa20bb07.
