### 2021-03-26 e7d80c7
Minor fix-moved opening curly braces
