### 2021-03-27 590430a
Updated documentation after changes in R files
