+++ man/get_biomass.Rd
@@ -12,7 +12,7 @@ get_biomass(
   new_eqtable = NULL,
   wna = 0.1,
   w95 = 500,
-  Nres = 10000
+  nres = 10000
 )
 }
 \arguments{
@@ -41,7 +41,7 @@ range. Default is 0.1}
 the value at which the sample-size-related weight reaches 95\% of its
 maximum value (max=1). Default is 500.}
 
-\item{Nres}{number of resampled values. Default is 1e4.}
+\item{nres}{number of resampled values. Default is 1e4.}
 }
 \value{
 A vector of class "numeric" of the same length as dbh, containing AGB
