### 2021-04-04 #436 @gonzalezeb
Hi Adam, thanks for checking in.
I am getting some errors after addressing some issues brought out by `lintr::lint_package()` and `goodpractice::gp()`, but my collaborator is not, so I am not sure if everything is good (at least not from my laptop).

When I run `devtools::test()` to test functionalities, I get that some tests are not running well. When I run goodpractice::goodpractice() on 3/26 I got "Supreme package..." but now I get a few errors. Maybe if you run some checks again you can let me know if you are still seeing issues..
