### 2021-04-05 #436 @adamhsparks
There is a note about UTF-8 strings as well that I'm seeing that you'll need to address as well.

```r
* checking data for non-ASCII characters ... NOTE
  Note: found 156 marked UTF-8 strings
```
