### 2021-04-05 #436 @gonzalezeb
Thanks @adamhsparks. Most of the suggestion after running `lint_package()` have already been applied, see for example [here](https://github.com/forestgeo/allodb/blob/master/R/est_params.R)  and [here](https://github.com/forestgeo/allodb/blob/master/R/new_equations.R). I am not sure why they don't show to you. I still need to work on some thought. Let me know if after installing the package again you still see those.
