### 2021-04-05 #436 @adamhsparks
Ah, sorry, I didn’t pay attention and realise you’d applied the changes to the repo yet. I’d need to refresh my local branch and rerun. I thought you were having issues identifying and applying fixes to these initial problems.

>
> On 5 Apr 2021, at 09:47, Erika Gonzalez-Akre ***@***.***> wrote:
>
> ﻿
> Thanks @adamhsparks. Most of the suggestion after running lint_package() have already been applied, see for example here and here. I am not sure why they don't show to you. I still need to work on some thought. Let me know if after installing the package again you still see those.
>
