### 2021-04-05 bcaa4d1
Update new_equations.R
Needed to change column names on kgc::climatezones dataframe to small letters, so it would take snake_style
