+++ R/new_equations.R
@@ -112,7 +112,7 @@ new_equations <- function(subset_taxa = "all",
     eq_jansen <- merge(eq_jansen, hallom, by = "genus")
     # substitute H by its DBH-based estimation
     to_merge <- eq_jansen[, c("hsub", "equation_allometry")]
-    eq_jansen$equation_allometry <- apply(to_merge, 1, function(X) {
+    eq_jansen$equation_allometry <- apply(to_merge, 1,  function(X) {
       gsub("\\(h", paste0("((", X[1], ")"), X[2])
     })
     # replace independent_variable column
@@ -270,8 +270,9 @@ new_equations <- function(subset_taxa = "all",
                       lat = new_coords[, 2])
     rcoords_eq <- round(coords_eq * 2 - 0.5) / 2 + 0.25
     ## extract koppen climate of every location
+    colnames(climatezones) <- c("lati", "longi", "cls")
     koppen_zones <- apply(rcoords_eq, 1, function(X) {
-      subset(kgc::climatezones, Lon == X[1] &  Lat == X[2])$Cls
+      subset(kgc::climatezones, longi == X[1] &  lati == X[2])$Cls
     })
     koppen_zones <- as.character(unlist(koppen_zones))
     if (length(koppen_zones) != nrow(rcoords_eq)) {
