### 2021-04-05 b8ac896
Change to suppressWarnings to snake_case
