### 2021-04-05 c00fb98
Changed to snake_case and reversed names for koppen...
climatezones datframe shouodl ramain as origical otherwise I get the message "No visible binding for global variable" if changing the colnames
