### 2021-04-05 8248efe
Changed to snake_case and reverse kppen climatezones..
colnames can't be changed
