### 2021-04-06 501188f
Reversed we to wE for kopmatrix
