### 2021-04-07 effe46f
changed we to wE in weight_allom : tests pass again
