### 2021-04-14 62f30be
Fixed local variable, now using it
To use the value saved into the variable the call to paste0() should not use quotes
