+++ R/imports.R
@@ -3,8 +3,8 @@ NULL
 
 # FIXME: Hack to use stringsAsFactors = FALSE everywhere, with all versions of R
 # Can we use tibble() instead? (We are using dplyr, and it exports tibble().)
-data_frame <- function(...) {
-  base::data_frame(..., stringsAsFactors = FALSE)
+data.frame <- function(...) {
+  base::data.frame(..., stringsAsFactors = FALSE)
 }
 
 globalVariables(
