+++ R/imports.R
@@ -3,8 +3,8 @@ NULL
 
 # FIXME: Hack to use stringsAsFactors = FALSE everywhere, with all versions of R
 # Can we use tibble() instead? (We are using dplyr, and it exports tibble().)
-data.frame <- function(...) {
-  base::data.frame(..., stringsAsFactors = FALSE)
+tibble <- function(...) {
+  tibble::tibble(..., stringsAsFactors = FALSE)
 }
 
 globalVariables(
