### 2021-04-19 #436 @gonzalezeb
Hi @adamhsparks. Most issues brought out by `lint` have been addressed, only two remain but I am little stuck on them. Maybe you can give me some guidance.
![image](https://user-images.githubusercontent.com/26872628/115269359-3ae81d00-a109-11eb-9f0f-101f3d4adfe3.png)

In `R/new_equations`, the function contains a lot of 'if' because it does a lot of sanity checks on the user-provided data. We could remove these, at the risk of having undetected problems in the user-provided allometries; and it seems that moving the 'ifs' to another function would make much sense.

In `R/weight_allom`, the problem is with "wE" which is the original col name of a matrix (koppenMatrix) been used.
