### 2021-04-23 #436 @adamhsparks
@gonzalezeb, can you please update your README badges for _allodb_ using `rodev::use_review_badge(436)` and add a NEWS.md file, `usethis::use_news_md()`?
