### 2021-04-23 #436 @adamhsparks
@gonzalezeb, can you also please address the errors that are encountered when running checks?

```r
==> devtools::check()

ℹ Updating allodb documentation
ℹ Loading allodb
Writing NAMESPACE
Writing NAMESPACE
── Building ────────────────────────────────────────────────────────── allodb ──
Setting env vars:
● CFLAGS    : -Wall -pedantic -fdiagnostics-color=always
● CXXFLAGS  : -Wall -pedantic -fdiagnostics-color=always
● CXX11FLAGS: -Wall -pedantic -fdiagnostics-color=always
────────────────────────────────────────────────────────────────────────────────
✓  checking for file ‘/Users/adamsparks/Development/GitHub/Reviews/allodb/DESCRIPTION’ ...
─  preparing ‘allodb’:
✓  checking DESCRIPTION meta-information ...
E  creating vignettes (5.3s)
   --- re-building ‘allodb-vignette.Rmd’ using rmarkdown
   Quitting from lines 81-87 (allodb-vignette.Rmd)
   Error: processing vignette 'allodb-vignette.Rmd' failed with diagnostics:
   arguments imply differing number of rows: 9991, 19
   --- failed re-building ‘allodb-vignette.Rmd’

   SUMMARY: processing the following file failed:
     ‘allodb-vignette.Rmd’

   Error: Vignette re-building failed.
   Execution halted

Error: 'col_ed' is not an exported object from 'namespace:cli'
Execution halted

Exited with status 1.
```

Once that's dealt with, I'll look for reviewers. 🙏
