### 2021-04-27 5a01a25
Update DESCRIPTION
didn't work with usethis:: trying without it
