### 2021-04-27 cb8f41b
Update DESCRIPTION
TRy using 'tibble' because use_package(tibble) was giving me this error:
Error: Failed to install 'allodb' from GitHub:
  Invalid comparison operator in dependency: use_package("tibble")
