+++ DESCRIPTION
@@ -30,8 +30,8 @@ Imports:
     data.table (>= 1.12.8),
     ggplot2 (>= 3.3.2),
     kgc (>= 1.0.0.2),
-    utils,
-    use_package("tibble")
+    tibble,
+    utils
 Suggests: 
     covr (>= 3.2.1),
     knitr (>= 1.21),
