### 2021-05-03 ec33472
add mention to tibble package and revert changes in resample_agb()
