### 2021-05-03 47fcbb7
Add NEWS.md
