+++ NEWS.md
@@ -1,0 +1,3 @@
+# allodb 0.0.0.9000
+
+* Added a `NEWS.md` file to track changes to the package.
