### 2021-05-03 2bb2732
Updated README as suggested by editor
same as creating NEWS (previous commit)
