### 2021-05-04 8ea4a5e
If using data.frame(), then use stringsAsFactors = FALSE
