+++ R/illustrate_allodb.R
@@ -61,9 +61,9 @@ illustrate_allodb <- function(genus,
   eq_info <- apply(equations[, c("equation_id", eqinfo)], 1,
                    function(x)
                      paste(x, collapse = " - "))
-  eq_info <- data.frame(equation_id = equations$equation_id,
+  eq_info <- data.frame(stringsAsFactors = FALSE, equation_id = equations$equation_id,
                         equation = eq_info)
-  dfcounts <- data.frame(table(equation_id = dfagb$equation_id))
+  dfcounts <- data.frame(stringsAsFactors = FALSE, table(equation_id = dfagb$equation_id))
   eq_info <- merge(eq_info, dfcounts, by = "equation_id")
   eq_info <- eq_info[order(eq_info$Freq, decreasing = TRUE), ]
   eq_info[(neq + 1):nrow(eq_info), "equation"] <- "other"
