### 2021-05-15 #436 @adamhsparks
Hi @gonzalezeb, I'm getting a NOTE when I run `R CMD check`.

```r
Note: found 156 marked UTF-8 strings
```

Can you check into this?

As it's just a note, not an error in the checks, I'll start looking for reviewers now.
