### 2021-05-20 #436 @jeffreyhanson
Hi @gonzalezeb, I just had a quick look through this package it looks really awesome! Great work!

Before I do a through review, I have some broad suggestions for the code/doucmentation that are relevant for multiple functions. I also noticed a few specific things that might be good to fix before I do the reivew too. I've listed them below - what do you think? I think most of them could be addressed by polishing the code/documentation a bit more? I'm also happy to conduct a through review with the current version of the package if you prefer - it will just a lot of small issues (e.g. this term in this file should have code formatting) that might be annoying for you to have to respond to?

1. The package website isn't working (https://forestgeo.github.io/allodb/). I think there might be a bug in the GitHub Actions code for generating the package website.
2. I feel like the function documentation could be improved by using code formatting (e.g. use ["\`character\` vector" instead of "character vector"](https://github.com/forestgeo/allodb/blob/master/R/get_biomass.R#L17)). This would help users understand that you are referring to a specific data type. This suggestion applies to multiple functions - what do you think?
3. I feel like the function documentation could be improved by using more precise terminology. For example, in the documentation for function parameters, I think it would be helpful to specify the exact class for arguments (e.g. using ["\`numeric\` vector" instead of "numerical vector", because "numerical" isn't a class in R](https://github.com/forestgeo/allodb/blob/master/R/get_biomass.R#L13)). Additionally, I think it would be helpful to include classes in the documentation for function outputs (i.e. in `@return` sections). For example, [this sentence](https://github.com/forestgeo/allodb/blob/master/R/resample_agb.R#L23-L26) currently says "A data frame of resampled DBHs and associated ..." and could be updated to say "A data frame (\`tibble::tibble()\` object) of resampled DBHs and associated ...". This suggestion applies to multiple functions - how does that sound?
4. It's not really the best practice to incude a non-R script file in the /R folder. Perhaps the the [`sysdata.rda`](https://github.com/forestgeo/allodb/blob/master/R/sysdata.rda) file could be moved to a more apropriate location?
5. I feel like the it might helpful to include some code to validate function arguments. This could help users understand why their code isn't working if they supply incorrect arguments (e.g. if they supply a `numeric` vector instead of a `character` vector). For example, [I think this argument validation code is fantastic and will be really useful for users](https://github.com/forestgeo/allodb/blob/master/R/resample_agb.R#L42-L44), and I wonder if you could add similar code to verify arguments to all parameters in each function? If you're interested in R packages to help with this, I personally use the [assertthat R package](https://github.com/hadley/assertthat) (e.g. [see this function](https://github.com/prioritizr/wdpar/blob/master/R/wdpa_read.R#L40-L51)) and I have heard good things about [checkmate R package](https://github.com/mllg/checkmate) too. What do you think?
6. Could you remove the unused Travis badge in REAMDE (because it seems that GitHub Actions is used for CI)?
7. The .buildignore folder seems to contain random R script and meeting notes. I don't see the benefit for keeping this in the code repository? Perhaps you could move them to a wiki, issues, or a seperate repo? Sorry if I'm missing something?
8. [This function](https://github.com/forestgeo/allodb/blob/ec33472e66d5ed8ac7e798bb38fa31e75fa018f3/R/resample_agb.R) (and maybe others?) over-ride the global state. This is not ideal because it could cause issues for users when they do not expect such changes. For example, imagine a user is trying to use this package within an analysis, and they manually use `set.seed()` to handle random number generation for simulations in various parts of their analysis. If this function uses `set.seed()` and does not revert it afterwards, then it could mean that the user's subsequent simulations are all based on the same seed even though they used `set.seed()` to change the seed in their code. If you're interested in suggestions on how to address this, I personally like to use the [withr R package](https://www.rdocumentation.org/packages/withr/versions/2.4.2/topics/with_seed) for stuff like this. For example, instead of using the [`set.seed()` function](https://github.com/forestgeo/allodb/blob/ec33472e66d5ed8ac7e798bb38fa31e75fa018f3/R/resample_agb.R#L79), you could use the [withr::with_seed()` function](https://www.rdocumentation.org/packages/withr/versions/2.4.2/topics/with_seed). Does that help?
9. `eval(parse(....))` is rarely a good idea unless you have no alternatives. This is because it can introduces bugs that are hard to track down and it can also cause security problems. [In this function](https://github.com/forestgeo/allodb/blob/ec33472e66d5ed8ac7e798bb38fa31e75fa018f3/R/resample_agb.R#L94), I wonder if you could you could use some combination of `expression` and `substitute` to acheive the same result?
10. I feel like it would be easier for users to use [this function](https://github.com/forestgeo/allodb/blob/master/R/get_biomass.R) (and maybe others too?) if the parameters were re-ordered such that all parameters without defaults came before those with defaults. For example, to run the `get_biomass()` function with default parameters and without using named arguments, I would have to do something like `get_biomasss(x, y, , z)` instead of get_biomass(x, y, z)` (to specify arguments for `dbh`, `genus`, and `coords`). But maybe there's a reason for this particular order for parameters? Sorry if I'm missing something obvious.
