### 2021-05-24 f77f125
Update README.md
 Remove unused Travis badge as suggested (GitHub Actions is used for CI)
