### 2021-05-24 5d4c675
Removed .buildignore folder
Kept in Erika's files, as if needed
