+++ .buildignore/issues/38_valid_eqn/38_valid_eqn.Rmd
@@ -1,57 +1,0 @@
----
-title: "Check if equations contain valid R code"
-output: github_document
----
-
-```{r setup, include=FALSE}
-knitr::opts_chunk$set(echo = TRUE)
-```
-
-```{r}
-library(allodb)
-library(tidyverse)
-```
-
-Data.
-
-```{r}
-data("equations", package = "allodb")
-eqn <- equations$equation_allometry
-```
-
-Try to evaluate each equation and catch errors.
-
-```{r}
-eval_toy <- function(text) {
-  toy_value <- 10
-  eval(parse(text = text), envir = list(dbh = toy_value))
-}
-
-contents <- eqn %>% 
-  map(safely(eval_toy)) %>% 
-  transpose()
-
-ok <- map_lgl(contents$error, is.null) 
-
-invalid <- tibble(
-  equation_allometry = eqn[!ok], 
-  messages = map_chr(contents$error[!ok], "message")
-)
-
-unique(invalid$messages)
-```
-
-Issue with 'dba'.
-
-```{r}
-filter(invalid, grepl("dba", messages))
-```
-
-Proposed solution: [#38](https://github.com/forestgeo/allodb/issues/38) and [#41](https://github.com/forestgeo/allodb/issues/38).
-
-Issue with 'BA'.
-
-```{r}
-filter(invalid, grepl("BA", messages))
-```
- How should we deal with this?
