+++ .buildignore/issues/41_shrubs/41_shrubs.md
@@ -1,13 +1,0 @@
-Issue 41: Calculate biomass of shrubs
-================
-
-Here I wrap discussions in issue 41 and try to come up with my next
-actions for developing code.
-
-  - Where is Valentine’s code again?
-
-  - Write expressive conditions to use in if statements to control flow:
-    
-      - `is_shrub()`.
-      - `depends_on_dbh()`.
-      - `depends_on_dba()`.

[…]
