### 2021-05-25 c44ad5b
changed `set.seed` with `withr::with_seed` in the `resample_agb` function, as suggested by @jeffreyhanson in issue #436
