### 2021-05-25 3015ade
changed `parse` into more explicit `str2lang`
