+++ R/resample_agb.R
@@ -7,9 +7,9 @@
 #'
 #' @param genus a character value, containing the genus (e.g. "Quercus") of the
 #'   tree.
+#' @param coords a numerical vector of length 2 with longitude and latitude.
 #' @param species a character value, containing the species (e.g. "rubra") of
 #'   the tree. Default is NULL, when no species identification is available.
-#' @param coords a numerical vector of length 2 with longitude and latitude.
 #' @param new_eqtable Optional. An equation table created with the
 #'   add_equation() function. Default is the original allodb equation table.
 #' @param wna this parameter is used in the weight_allom function to determine
@@ -33,8 +33,8 @@
 #' )
 #'
 resample_agb <- function(genus,
-                         species = NULL,
                          coords,
+                         species = NULL,
                          new_eqtable = NULL,
                          wna = 0.1,
                          w95 = 500,
