+++ R/get_biomass.R
@@ -14,12 +14,12 @@
 #'   measurements, in cm.
 #' @param genus a character vector (same length as dbh), containing the genus
 #'   (e.g. "Quercus") of each tree.
-#' @param species a character vector (same length as dbh), containing the
-#'   species (e.g. "rubra")  of each tree. Default is NULL, when no species
-#'   identification is available.
 #' @param coords a numerical vector of length 2 with longitude and latitude (if
 #'   all trees were measured in the same location) or a matrix with 2 numerical
 #'   columns giving the coordinates of each tree.
+#' @param species a character vector (same length as dbh), containing the
+#'   species (e.g. "rubra")  of each tree. Default is NULL, when no species
+#'   identification is available.
 #' @param new_eqtable Optional. An equation table created with the
 #'   new_equations() function.
 #' @param wna this parameter is used in the weight_allom function to determine
@@ -48,8 +48,8 @@
 #'
 get_biomass <- function(dbh,
                         genus,
-                        species = NULL,
                         coords,
+                        species = NULL,
                         new_eqtable = NULL,
                         wna = 0.1,
                         w95 = 500,
