+++ man/resample_agb.Rd
@@ -6,8 +6,8 @@
 \usage{
 resample_agb(
   genus,
-  species = NULL,
   coords,
+  species = NULL,
   new_eqtable = NULL,
   wna = 0.1,
   w95 = 500,
@@ -18,10 +18,10 @@ resample_agb(
 \item{genus}{a character value, containing the genus (e.g. "Quercus") of the
 tree.}
 
+\item{coords}{a numerical vector of length 2 with longitude and latitude.}
+
 \item{species}{a character value, containing the species (e.g. "rubra") of
 the tree. Default is NULL, when no species identification is available.}
-
-\item{coords}{a numerical vector of length 2 with longitude and latitude.}
 
 \item{new_eqtable}{Optional. An equation table created with the
 add_equation() function. Default is the original allodb equation table.}
