### 2021-06-02 forestgeo/allodb#147 @maurolepore
Backport str2lang()?
The function str2lang() was released in R 3.6.0 so the package fail for users of R < 3.6.0. You can see this on CI (notice [the workflow fails with ubuntu and R 3.5.0](https://github.com/forestgeo/allodb/actions/runs/900235963)).

https://github.com/r-lib/backports#backports-for-r-versions-prior-to-360

I'm not familiar with that funciton so not sure how easy/hard is to
re-implement. Alternatively we could backport it with the backports
package: https://github.com/r-lib/backports#in-packages﻿
