+++ DESCRIPTION
@@ -22,7 +22,7 @@ Authors@R:
 Description: Tool to standardize and simplify the tree biomass
     estimation process across globally distributed extratropical forests.
 License: GPL-3
-URL: https://github.com/forestgeo/allodb
+URL: https://forestgeo.github.io/allodb, https://github.com/forestgeo/allodb
 BugReports: https://github.com/forestgeo/allodb/issues
 Depends: 
     R (>= 3.5.0)
