+++ _pkgdown.yml
@@ -1,0 +1,1 @@
+url: https://forestgeo.github.io/allodb
