### 2021-06-03 forestgeo/allodb#147 @cpiponiot
Ok, I'll change the function back to `parse`
