### 2021-06-03 66c5e1a
change str2lang() back into parse() to be compatible with R versions < 3.6.0 (issue #147)
