### 2021-06-14 #436 @jstillh
Hi @gonzalezeb
Here is my review. I really appreciate the package and the work you & the co-authors have put into developing it.
I think the package will be useful for a lot (if not every) forest ecologist...  I reviewed the package looking at it as a regular user - sorry if i misunderstood some of the components of your package.

-   Besides Kristina Anderson Texeira and myself being listed as co-authors on a paper with 100+ authors i don’t have any    professional relationship with the authors of the package
-   [x] As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to
    review this work (If you are unsure whether you are in conflict,  please speak to your editor *before* starting your review).

-   [x] **Examples** (that run successfully locally) for all exported functions

-   [x] Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer (“rev” role) in the package   DESCRIPTION file.

The authors provide a package allowing to calculate biomass for individual trees of extratropical tree and shrub species. A package including biomass equations for extratropical species is a useful and  timely complement to the `BIOMASS` package by Chave et al. for tropical species. The authors apparently developed the package focusing on the plots of the ForestGEO network. Given that the authors aim at providing a package complementing the `BIOMASS` package, i reviewed the functionality of the package not only for the use on single/multiple sites but for the use along latitudinal gradients and focused on European tree species that i am familiar with.
Although i appreciate the general idea of the package and the large effort that the authors put into its development, i have some concerns regarding the concept applied when calculating biomass and some general comments on code and documentation that i feel should be addressed:

#### Conceptual

The core function `get_biomass` provides the user with one estimate of aboveground biomass based on the diameter at breast height (dbh) of a tree. This function then applies one or several of the biomass equations based on climatic similarity and similarity between species. The estimate is based on a large set of biomass-equations compiled from a wide range of sources - i very much appreciate the work the authors have put into the compilation of this dataset.
I like the general framework of this package and the option to add additional equations via the function `new_equations` but i have concerns regarding the workflow implemented in the package. These mainly focus on the following points:

1.  *Documentation of selection and weighting*
    The function `get_biomass` returns a `vector` with an estimate of   the biomass for each tree. The calculation of this value is somewhat   opaque as long as the user does not actively change some of the arguments (which by default do not need to be provided) of the  `get_biomass` function. Even if species / genus specific equations  are available, this function includes additional, less specific equations. These are given a certain weight based on similarities using the `weight_allom` function and might contribute little to the final estimate. However, this is not obvious to the user, as the package only provides one value for each tree. I would appreciate if the function `get_biomass` would not only return a vector but more metadata on the steps that resulted in the calculation of the final result. This could be achieved by returning a list containing the final values, all values with the associated weights, the output of `new_eqtable` documenting the equations used and the output of `weight_allom`.
    I understand that the function `illustrate_allodb` should give the user an impression on the results of the resampling - but it does  not provide the user with all information (e.g., no information on  weights) and it requires the user to run an additional function.

2.  *Inclusion of equations based on climatic simlilarity*
    To check for climatic similarity, the package makes use of the   Koeppen classification. This is a feasible approach - but it has
    some drawbacks in the implementation: Based on the Koeppen    classification of the site (provided through the `coords` argument),    a certain set of equations is used. This can result in the exclusion/ downweighting even of species specific biomass equations. The  illustrative example below provides biomass estimates for *Fagus  sylvatica* along a latitudinal gradient from Northern Italy to  Northern Germany and the output of `illustrate_allodb` for two sites used in this latitudinal gradient. Obviously a completely different  set of equations is used - in the second site, no species specific
    equations are used. (Btw: there seems to be a typo, i guess it should be *Populus* instead of *Pupulus*.)
    The differences between the sites observed here are mostly driven by a change in the Koeppen zones and a different set of equations applied due to the change in Koeppen zones. I am not convinced that
    this makes ecologically sense. Additionally, the  Koeppen-classification for specific equations seems to be based on
    the value for longitude and latitude provided in the original publication of each biomass equation. However, for equations
    covering a wider range (e.g., the equations published in Forrester et al. 2017), the authors seem to have extracted the Koeppen value for the mean of both the lat and long range provided. This may result in a very specific classification not representative for the species and a subsequent down-weighting of this equation when applied in the midst of the species’s distribution range: For the  ForestGeo Plot in Zofin, which is mainly covered with beech and lies in the center of the distribution range of this species, this results in a down-weighting of the equation provided in Forrester et al. 2017 as the value `wE` provided in the dataset for this specific  combination in `koppenMatrix` is 0.7. I did not check if other equations have a higher weight, but i think this issue should be  addressed by the authors.
    If the Koeppen classification is not included in the dataset  `koeppenMatrix` or the specific combination if Koeppen zones is
    included with a weight of 0, this results in the exclusion of a   specific equation and a subsequent error message (if only one
    equation is provided) or the inclusion of other, not species  specific equations, see my comments further down.

In general, the weighting concept of the different biomass equations should be more transparent for the package user and probably requires conceptual changes.

``` r
data <- allodb::get_biomass(rep(50, 41), rep("Fagus", 41), rep("sylvatica", 41), coords = as.matrix(cbind(rep(10, 41), seq(44, 54, 0.25))))
rcoords_site <- round(as.matrix(cbind(rep(10, 41), seq(44, 54, 0.25))) * 2 - 0.5)/2 + 0.25
koeppen <- apply(rcoords_site, 1, function(xk) {
  subset(kgc::climatezones, Lon == xk[1] & Lat == xk[2])$Cls
})
plot(seq(44, 54, 0.25), data, xlab = "lat", ylab = "biomass [kg]",  ylim = c(0, 3000), pch = 21, cex = 1.5, bg = koeppen)

![unnamed-chunk-1-1](https://user-images.githubusercontent.com/31648011/121906533-7da22c00-cd2b-11eb-97fe-fe8d6a432e5f.png)

allodb::illustrate_allodb("Fagus" , coords = c(10, 44), "sylvatica")
```
![unnamed-chunk-1-1](https://user-images.githubusercontent.com/31648011/121906533-7da22c00-cd2b-11eb-97fe-fe8d6a432e5f.png)

``` r
allodb::illustrate_allodb("Fagus" , coords = c(10, 46), "sylvatica")
```

``` r
data("koppenMatrix", package = "allodb")
data("sites_info", package = "allodb")

koppenMatrix$wE[koppenMatrix$zone1 == sites_info$koppen[sites_info$site == "zofin"] & koppenMatrix$zone2 == "Csb"]
```

    ## [1] 0.7

#### Code and documentation

*Use of `eval(parse())`*
As @jeffreyhanson already pointed out [here](https://github.com/ropensci/software-review/issues/436#issuecomment-844578387), the use of `eval(parse())` is not good practice for the reasons he laid out nicely in his comment. You seem to have replaced the `eval(parse())` in [this function](https://github.com/forestgeo/allodb/blob/master/R/resample_agb.R) but then changed it back according to [this issue](https://github.com/forestgeo/allodb/issues/147) because the `str2lang` that you used to replace `eval(parse())` is not available for users with R &lt; 3.6.0.
I was not aware of the `str2lang` function and i am not convinced that using this function will solve the issue raised by @jeffreyhanson: The issue with tracing bugs will still exist even when changing to the `str2lang` function and the general security concerns associated with the use `eval(parse())` can probably not be solved with replacing the `parse` function by `str2lang` as `str2lang` is only a [special case of `parse`](https://stat.ethz.ch/R-manual/R-devel/library/base/html/parse.html).
The concerns raised are associated with the **combination** of `eval(parse())` and not with the use `parse()`. (btw: when you removed `str2lang` in [this function](https://github.com/forestgeo/allodb/blob/master/R/resample_agb.R), you only removed it on line 93 but not within the `else` statement on line 101).
I think working along the lines laid out by @jeffreyhanson (i.e. using `expression` and `substitute`) would be a good option. Another option would be to create internal functions for all forms of the equations in `dfequation` and then apply these.

*Documentation*
I think the vignette should be improved. It would be great if you could provide some more information to the user especially on the estimation of the parameters in `est_param` and what effect the weighting in `weight_allom` has on the result.
As @jeffreyhanson already pointed out [here](https://github.com/ropensci/software-review/issues/436#issuecomment-844578387), consistently using code formatting throughout the documentation would be great (e.g., “numeric” in the description of the `resample_agb` function could be change to `numeric`).

*Naming conventions*
The use of the package could be improved if you consistently order the arguments and stick to one naming convention throughout the package. `sitespecies` does not follow the naming convention you used for the datasets in the package - all other datasets are named following the convention name1\_name2. This applies for the dataset `koeppenMatrix` as well.

*Performance*
The calculation of the biomass is relatively slow if for every tree a unique coordinate is provided - i guess this is caused by the subset operations performed in the `weight_allom` function. Given that a user might provide individual coordinates for each tree on a site and the coarse resolution of the koeppen classification, it might be an option to first check if the coordinates vary in such a way that these need to be subset individually or if a single coordinate can be used.

Such performance considerations are necessary, as the calculations are rather slow at the moment. For example, the biomass calculation for 100 trees with tree level coordinates require around 9s.

``` r
system.time(allodb::get_biomass(dbh = c(1:100), genus = rep("Fagus", 100)
                                , species = rep("sylvatica", 100)
                                , coords = as.matrix(cbind(rep(10, 100), sample(seq(48.001, 48.1, 0.0001), 100)))))
```

    ##    8.81    0.03    8.86

#### Specific comments

`get_biomass`
The `get_biomass` function creates a biomass estimate based on `dbh`, `genus`, `coords` and `species`. It allows the user to provide a selection of the biomass equations via the argument `new_eqtable`. This seems to be an important feature of this function. However, if the values provided in the column `koppen` do not appear in `kgc::climatezones`, the function throws an error and the error message does not point the user to the reason of the error - see my specific comment on the `weight_allom` function below. Given the coarse resolution of the `kgc::climatezones` dataset of 0.5 degrees, this might be problematic especially in mountainous regions with a large altitudinal gradient. Additionally, this is an issue which might prevent users from applying equations that could be applied for the respective  region - see my general comment above.

``` r
equations_to_use <- allodb::new_equations(subset_taxa = "Fagus")

allodb::get_biomass(dbh = 20
                    , genus =  "Fagus"
                    , species = "sylvatica"
                    , coords = c(10, 46)
                    , new_eqtable = equations_to_use)
```

    ## Error in if (any(nEQ <- vNms != make.names(vNms))) vNms[nEQ] <- paste0("`", : missing value where TRUE/FALSE needed

``` r
allodb::get_biomass(dbh = 20
                    , genus =  "Fagus"
                    , species = "sylvatica"
                    , coords = c(10, 46))
```

    ## [1] 217.0154

`weight_allom`

The `weight_allom` function creates weights for all biomass functions which will later be used in the calculation of the biomass.

Beyond the concerns i already mentioned above, i only have a minor comment on this function: The `kgc::climatezones` dataset does not cover large bodies of water. The values in `dfequation$wE` will therefore be `-Inf` for all `coords` that are not covered by `kgc::climatezones` The `weight_allom` function therefore returns `-Inf` causing the function to stop without a easily  understandable error message. I would appreciate if you either add an error message like `if(dfequation$wE == -Inf){stop('Error message')}` or if you would just flag the respective values so they are omitted in later calculations.

`new_equations`
The function `new equations` allows the user to a) subset to a specific set of equations for one or several taxa and b) to add additional equations. I appreciate the latter option, as this adds a lot of flexibility to the package.
