### 2021-06-18 #436 @jeffreyhanson
- [x] **Examples** (that run successfully locally) for all exported functions

---

The allodb R package provides a framework for estimating tree biomass based on allometric equations. Overall, my impression is that the package is very well designed. I think the authors have done a fantastic job of providing a highly flexible interface --- allowing users to customize calculations as needed --- that is easy to use. Since I have zero expertise in forestry science, I cannot comment on the novelty, scope, or methodological correctness of the package. As such, I have limited my review to focus on code and documentation.

I previously posted several suggestions for improving the package. Although I am glad that some of these suggestions were used to help improve the package, I would appreciate hearing why some of these suggestions were not incorporated? It's entirely possible that these suggestions were irrelevant, infeasible, or undesireable. For example, is there some particular benefit for not describing classes precisely, or not validating user inputs of which I am unaware? Below I've listed some specific comments. Please let me know if anything I have written is unclear, irrelevant, or incorrect?

- I mentioned earlier that I thought it would be helpful to include some code to validate function arguments. From what I can tell, the code hasn't been updated to include such validation - is that correct? I'm sorry, was my previous comment unclear, or did I misunderstand something? To help show what I mean, here's some example code that could be used to validate arguments for [the `est_params()` function](https://github.com/forestgeo/allodb/blob/master/R/est_params.R#L40-L48). Please note that I'm not saying you have to use this specific code, I'm just trying to provide an example of what the code could look like (e.g. you might prefer to use the checkmate R package instead):
    ```
    est_params <- function(genus,
                           coords,
                           species = NULL,
                           new_eqtable = NULL,
                           wna = 0.1,
                           w95 = 500,
                           nres = 1e4
    ) {
    # assert arguments are valid
    assertthat::assert_that(
      ## genus
      is.character(genus),
      assertthat::noNA(genus),
      ## coords
      is.numeric(coords),
      assertthat::noNA(coords),
      length(coords) == 2
      ### ensure valid longitude
      isTRUE(coords[1] >= -180),
      isTRUE(coords[1] <= 180),
      ### ensure valid latitude
      isTRUE(coords[2] >= -90),
      isTRUE(coords[2] <= 90),
      ## species
      inhrerits(species, c("character", "NULL")),
      ## new_eqtable
      inhrerits(new_eqtable, c("data.frame", "NULL")),
      ## wna
      assertthat::is.number(wna),
      assertthat::noNA(wna),
      ## w95
      assertthat::is.number(w95),
      assertthat::noNA(w95),
      ## nres
      assertthat::is.count(nres),
      assertthat::noNA(nres)
    )
    ```

- I previously suggested that it would be good to avoid using `eval(parse(...))` if possible. Looking through the commit history, it seems like an alternative function was attempted (i.e. `str2lang(...)`) but this was later removed to avoid compatibility issues with older R versions. Thanks for looking into this! I appreciate these efforts of trying to get it working. I was trying to think of alternative strategies to avoid  `eval(parse(...))`, but I think the over-arching issue is that some equivalent is needed because the equations are (ultimately) stored as `character` objects (e.g. something like `"log(dbh) * 5"`). I suppose some alternatives would be to update the `equations` data frame to replace the `character` objects with `function` objects, or replacing the `equations` data frame with a `list` of purpose-built equation objects that can be used to perform the calculations. However, these seem needlessly complex compared to the current implementation? Maybe this could be one instance where `eval(parse(...))` is not bad idea?

    To help avoid potential issues with `eval(parse(...))`, maybe it would be worth (1) writing a function that can validate if a given `character` is a valid equation or not (it could give a useful error message if the function is not valid), (2) adding this validation function to the other functions that accept equation tables as inputs, and (3) adding unit tests to validate all built-in equation (i.e. specified in the `equation` table) using this new function? I'm sorry that I can't be of more help.

- The package contains functions that output different data frame classes. For example, the `resample_agb()` function outputs a `tibble` (well, `tbl_df` object), and the `est_params()` function outputs a `data.table` object. Since these functions output different classes, this means that the user will need to keep remember exactly what data frame class is output from each function -- increasing the cognitive load required to use the package. In other words, this behavior breaks the ["law of least surprise"](https://en.wikipedia.org/wiki/Principle_of_least_astonishment). Is there a particular reason for this? I'm sorry if I'm missing something obvious?

    To help make the package easier to use, may I suggest picking one data frame class (e.g. `data.frame`, `tibble` or `data.table`) and using that as the basis for all user-facing functions? This way the user can simply remember something like "The allodb works with `data.table` objects, so I need to make sure my input data are in that format, and convert any outputs from that format as needed". Indeed, if there are certain reasons for using certain data frame classes in certain functions (such as for performance), perhaps different classes could be used internally (e.g. a function gets a `data.frame` as input, converts it to a `data.table` for performance, does some processing, and then outputs a `data.frame`)? What do you think? I'm just trying to help suggest ways to make the package easier to use.

- The [`nls()` function used in `est_params()`](https://github.com/forestgeo/allodb/blob/master/R/est_params.R) is exceptionally useful for fitting nonlinear models. Since it can experience convergence issues (e.g. see [here](https://r.789695.n4.nabble.com/nls-convergence-problem-td805934.html), [here](https://r.789695.n4.nabble.com/suggestions-for-nls-error-false-convergence-td790026.html), [here](https://r.789695.n4.nabble.com/avoiding-termination-of-nls-given-convergence-failure-td886557.html), and [here](https://r.789695.n4.nabble.com/nls-singular-convergeance-td3900647.html)), it is sometimes necessary to customize the default settings to obtain reliable outputs (e.g. change optimizer, increase number of iterations). Given this, would it be possible to update the `est_params()` function so that users can customize the optimization parameters (e.g. specify the `algorithm` and `control` parameters for `nls()`, see [here for details](https://www.rdocumentation.org/packages/stats/versions/3.6.2/topics/nls))? It might also be useful if the function could somehow output information on the performance of the models too (e.g. model goodness of fit statistics)? But maybe I'm missing something and these concerns are not relevant here? What do you think?

- I mentioned earlier that I thought the documentation could be improved by using more precise terminology to describe function parameters and outputs. For example, using ["\`numeric\` vector" instead of "numerical vector", because "numerical" isn't a class in R](https://github.com/forestgeo/allodb/blob/master/R/get_biomass.R#L13)). Also, specifying exactly what data frame class (e.g. `data.frame` or `data.table`) when referring to R objects (e.g. [here](https://github.com/forestgeo/allodb/blob/master/R/resample_agb.R#L23-L26)). From what I can tell, the documentation hasn't been updated based on this suggestion - is that right? I'm sorry if my previous comment wasn't clear? Or maybe I'm misunderstanding something? If so, could you please explain why the current terminology is preferable?

- It would be difficult for me to understand exactly how the `get_biomass`, `est_params`, and `resample_agb` functions work based only on reading the documentation. I wonder if it is possible to provide more details on the underlying calculations in a Details section? This suggestion may not be relevant though, because I might lack the required experience in forestry science to understand the documentation in its current form?

- I mentioned earlier that I thought the documentation could be improved by using code tags to specify that certain words correspond to object classes. From what I can tell, it seems these tags are still missing from a few places (e.g. [like "NULL" here](https://github.com/forestgeo/allodb/blob/master/R/resample_agb.R#L12)). Could you please double check the documentation and add in any missing tags? Also, when documenting character objects, I would suggest including quotation marks (e.g. using `"all"` instead of `all` [here](https://github.com/forestgeo/allodb/blob/master/R/new_equations.R#L8)) so it's clear the documentation is referring to a character object.

- Have you thought about including more links to help the user navigate the documentation (e.g. [here where the documentation refers the reader to another manual entry](https://github.com/forestgeo/allodb/blob/master/R/data.R#L63), or [here where the documentation refers to another function](https://github.com/forestgeo/allodb/blob/master/R/get_biomass.R#L28))? Perhaps the `@seealso` section could also be used to help users navigate between related functions?

- Have you considered using equation commands (e.g. `\eqn{}`, `deqn{}`) in the documentation to format equations [e.g. [such as this equation](https://github.com/forestgeo/allodb/blob/master/R/est_params.R#L27); see here for example](http://www.deanbodenham.com/learn/r-package-documentation.html)?

- It might be helpful to use the `\pkg{}` command when referring to the name of packages in the documentation (e.g. such as `\pkg{allodb}` in [here](https://github.com/forestgeo/allodb/blob/master/R/resample_agb.R#L14))?

- Is it possible to explicitly state the type of object in the documentation for each dataset? I appreciate that the documentation says that the objects are a ["data frame"](https://github.com/forestgeo/allodb/blob/master/R/data.R#L12), but it would be useful to precise what implementation (e.g. `data.frame`, `data.table`, `table`, `tibble` objects are all of "data frame" objects).

- It might be helpful to add example sections to the documentation for the built-in datsets so that the reader can see a preview of each dataset. What do you think? For example, in [data.R](https://github.com/forestgeo/allodb/blob/master/R/data.R#L65-L66), you could insert some code like this:
    ```
    #' @examples
    #' # preview the dataset
    #' print(head(equations))
    #'
    ```

- I also noticed some potential typos in the documentation: [here](https://github.com/forestgeo/allodb/blob/master/R/data.R#L241), [here](https://github.com/forestgeo/allodb/blob/master/R/resample_agb.R#L6).

#### Built-in datsets

- I noticed that some of these objects are `tibble` objects (e.g. `equations`) and others are `data.frame` objects (e.g. `genus_family`). I was wondering if it would be useful to have all built-in datasets represented using the same class (e.g. all as `tibble` objects)? This is because `tibble` and `data.frame` objects have markedly different behaviors -- even though a `tibble` inherits from a `data.frame` -- and [these differences can cause errors when the user/code expects a different behavior](https://stat.ethz.ch/pipermail/r-package-devel/2017q3/001896.html). This is related to the "law of least surprise" that I mentioned earlier.

#### Minor suggestions

- You could replace `seq_len(length(...))` with `seq_along(...)` [here](https://github.com/forestgeo/allodb/blob/master/R/get_biomass.R#L79).

- Perhaps `rep.int()` could be useful [here](https://github.com/forestgeo/allodb/blob/master/R/resample_agb.R#L105-L107). For example, would this provide the same output: `equation_id <- rep.int(dfsub$equation_id, dfsub$resample)`?

- I think you could remove the `library(allodb)` from each test script (e.g. [this line](https://github.com/forestgeo/allodb/blob/master/tests/testthat/test-get_biomass.R#L1))?
