### 2021-06-28 #436 @gonzalezeb
@jstillh and @jeffreyhanson,  thank you so much for your thoughtful and kind reviews. Your suggestions will improve the package 100%. I will contact you if we have questions, we truly appreciate your time on this.
