### 2021-07-07 dc4b8fa
Fixed minor typos
