### 2021-07-16 #436 @gonzalezeb
Hi @adamhsparks. Sorry, I am a little behind with this and have solved only minor issues. There are some comments on the conceptual aspects of the package (i.e.  [here](https://github.com/ropensci/software-review/issues/436#:~:text=Inclusion%20of%20equations%20based%20on%20climatic%20simlilarity)) where I get the point but changing our approach will require some (lot's!) of work, so I am giving it a careful thought.  How should I tackle issues like those?
