+++ man/resample_agb.Rd
@@ -45,7 +45,7 @@ by the \code{weight_allom} function.
 After attributing a weight to each equation in allodb using the weight_allom
 function, equations are then resampled within their original DBH range using
 \code{resample_agb}: the number of resampled values for each equation is
-proportional to its weight. It creates S3 objects of class "numeric".
+proportional to its weight .It creates S3 objects of class "numeric".
 }
 \examples{
 resample_agb(
