### 2021-07-26 7305e72
weight_allom now deals with coordinates that are not associated with any koppen climate zone
