### 2021-07-26 cf3e14e
weight column `wE` to lower case `we` (more consistent with other column names)
