### 2021-07-26 56b2d0a
consistency checks of the dbh and the coordinates provided by the user in the get_biomass function
