### 2021-07-26 31286d3
Update equations.Rd
