+++ eval-parse.Rmd
@@ -47,14 +47,16 @@ to the current implementation? Maybe this could be one instance where
 * We are now testing the equations are correct or will trigger an informative
 error message.
 
-TODO: LET ME KNOW IF YOU NEED HELP. BELOW I SHOW A GIST OF
-WHAT A `validate_equation()` MIGHT LOOK LIKE.
+TODO: LET ME KNOW IF YOU NEED HELP. BELOW I SHOW A GIST OF WHAT A
+`validate_equation()` MIGHT LOOK LIKE.
 
 ## Unsafe
 
-[Online we found](https://stackoverflow.com/questions/13649979/what-specifically-are-the-dangers-of-evalparse) that parsing text would be a risk if we hosted an application
-on a server we own and allow users to input arbitrary text. A malicious user 
-could damage our server.
+[Online we
+found](https://stackoverflow.com/questions/13649979/what-specifically-are-the-dangers-of-evalparse)
+that parsing text would be a risk if we hosted an application on a server we own
+and allow users to input arbitrary text. A malicious user could damage our
+server.
 
 But this is not our case. The allodb package is to be used as an analysis tool,
 directly on the users's computers. And the input text comes mainly not from
