### 2021-07-26 4378dd0
Polish docs
It's best to not hard code the size of an object. The resulting
documentation will automatically add such information.
