+++ man/sitespecies.Rd
@@ -6,18 +6,7 @@
 \alias{sitespecies_metadata}
 \title{Tables of sites and tree species used in allodb and associated metadata.}
 \format{
-A data frame with 1114 rows and 10 variables:\preformatted{[,1]	site	character
-[,2]	family	character
-[,3]	genus	character
-[,4]	species	character
-[,5]	variety	character
-[,6]	subspecies	character
-[,7]	latin_name	character
-[,8]	species_code	character
-[,9]	life_form	character
-[,10]	notes	character
-
-}
+An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1113 rows and 10 columns.
 
 An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 10 rows and 4 columns.
 }
@@ -28,11 +17,14 @@ sitespecies_metadata
 }
 \description{
 \itemize{
-\item \code{sitespecies}: Table of extratropical ForestGEO sites
-in allodb (n=24) and their tree species.
-\item \code{sitespecies_metadata}: Metadata for \code{sitespecies}
-table.
+\item \code{sitespecies}: Table of extratropical ForestGEO sites in allodb (n=24) and
+their tree species.
+\item \code{sitespecies_metadata}: Metadata for \code{sitespecies} table.
 }
+}
+\examples{
+sitespecies
+sitespecies_metadata
 }
 \seealso{
 Other database datasets: 
