+++ man/missing_values.Rd
@@ -3,22 +3,19 @@
 \docType{data}
 \name{missing_values}
 \alias{missing_values}
-\title{A table with explanations of missing values codes}
+\title{Explanations of missing values codes}
 \format{
-A small data frame with 4 rows and 3 variables:\preformatted{	[,1]	Code	character
-	[,2]	Definition	character
-	[,3]	Description	character
-
-}
+An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 4 rows and 3 columns.
 }
 \usage{
 missing_values
 }
 \description{
-A list of codes used to indicate missing information in equation table:
-\itemize{
-\item \code{missing_values}: Table with explanations of missing values codes.
+Explanation of the codes used to indicate missing information in equation
+table.
 }
+\examples{
+missing_values
 }
 \seealso{
 Other database datasets: 
