### 2021-07-27 407352d
Ensure all functions return tibbles
