+++ man/gymno_genus.Rd
@@ -5,7 +5,7 @@
 \alias{gymno_genus}
 \title{Gymnosperms identified in selected ForestGEO sites}
 \format{
-An object of class \code{data.table} (inherits from \code{data.frame}) with 95 rows and 3 columns.
+An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 95 rows and 3 columns.
 }
 \usage{
 gymno_genus
