### 2021-07-28 692d79f
Comments reviewed
@maurolepore The statement on lines ~59-61 is correct.

Would you implement 'validate_equation()'? I think you argument is pretty valid
