### 2021-08-01 467e06e
New validate_equations()
