### 2021-08-01 4cb5899
New validate_equations()
