### 2021-08-02 #436 @gonzalezeb
Hi @adamhsparks, thanks for checking. We have made some substantial progress but I still need to work in some minor stuff. The plan is to finish this before Sep 15 as we want to submit our revised manuscript to MEE also by then.  Will that be ok with you all?
