+++ NAMESPACE
@@ -5,6 +5,5 @@ export(get_biomass)
 export(illustrate_allodb)
 export(new_equations)
 export(resample_agb)
-export(validate_equations)
 export(weight_allom)
 importFrom(utils,data)
