### 2021-08-04 b6b24b9
Remove str2lang() in favor or parse(text = )
