### 2021-08-04 d1c0cd1
Revert "Use R >= 3.6"
This reverts commit 1660cd15d7b97d60472e0a61b58487c1a1740ec3.
