### 2021-08-10 c41991c
improve documentation of selection and weighting
