+++ R/get_biomass.R
@@ -29,6 +29,27 @@
 #'   maximum value (max=1). Default is 500.
 #' @param nres number of resampled values. Default is 1e4.
 #'
+#' @details `allodb` estimates AGB by calibrating a new allometric equation for
+#'   each taxon (arguments `genus` and  `species`) and location (argument
+#'   `coords`) in the user-provided census data. The new allometric equation is
+#'   based on a set of allometric equations that can be customized using the
+#'   `new_eqtable` argument. Each equation is then given a weight with the
+#'   `weight_allom` function, based on: 1) its original sample size (numbers of
+#'   trees used to develop a given allometry), 2) its climatic similarity with
+#'   the target location, and 3) its taxonomic similarity with the target taxon
+#'   (see documentation of the `weight_allom` function). The final weight
+#'   attributed to each equation is the product of those three weights.
+#'   Equations are then resampled with the function `resample_agb`: the number
+#'   of samples per equation is proportional to its weight, and the total number
+#'   of samples is provided by the argument `nres`. The resampling is done by
+#'   drawing DBH values from a uniform distribution on the DBH range of the
+#'   equation, and estimating the AGB with the equation. The couples of values
+#'   (DBH, AGB) obtained are then used in the function `est_params` to calibrate
+#'   a new allometric equation, by applying a linear regression to the
+#'   log-transformed data. The parameters of the new allometric equations are
+#'   then used in the `get_biomass` function by back-transforming the AGB
+#'   predictions based on the user-provided DBHs.
+#'
 #' @return A vector of class "numeric" of the same length as dbh, containing AGB
 #'   value (in kg) for every stem
 #'
