### 2021-08-19 3c43a83
Changed numerical vector to numeric and added class to wna and w95
