+++ R/est_params.R
@@ -7,7 +7,7 @@
 #' and location in question are given a higher weight in this process.
 #' @param genus a character vector, containing the genus (e.g. "Quercus") of
 #'   each tree.
-#' @param coords a numerical vector of length 2 with longitude and latitude (if
+#' @param coords a numeric vector of length 2 with longitude and latitude (if
 #'   all trees were measured in the same location) or a matrix with 2 numerical
 #'   columns giving the coordinates of each tree.
 #' @param species a character vector (same length as genus), containing the
@@ -15,10 +15,10 @@
 #'   identification is available.
 #' @param new_eqtable Optional. An equation table created with the
 #'   add_equation() function. Default is the compiled allodb equation table.
-#' @param wna this parameter is used in the weight_allom() function to determine
+#' @param wna a numeric vector, this parameter is used in the weight_allom() function to determine
 #'   the dbh-related and sample-size related weights attributed to equations
 #'   without a specified dbh range or sample size, respectively. Default is 0.1
-#' @param w95 this parameter is used in the weight_allom() function to determine
+#' @param w95 a numeric vector, this parameter is used in the weight_allom() function to determine
 #'   the value at which the sample-size-related weight reaches 95% of its
 #'   maximum value (max=1). Default is 500.
 #' @param nres number of resampled values. Default is 1e4.
