+++ R/get_biomass.R
@@ -9,11 +9,11 @@
 #' The function can run into some memory problems when used on large datasets
 #' (usually several hundred thousand observations).
 #'
-#' @param dbh a numerical vector containing tree diameter at breast height (dbh)
+#' @param dbh a numeric vector containing tree diameter at breast height (dbh)
 #'   measurements, in cm.
 #' @param genus a character vector (same length as dbh), containing the genus
 #'   (e.g. "Quercus") of each tree.
-#' @param coords a numerical vector of length 2 with longitude and latitude (if
+#' @param coords a numeric vector of length 2 with longitude and latitude (if
 #'   all trees were measured in the same location) or a matrix with 2 numerical
 #'   columns giving the coordinates of each tree.
 #' @param species a character vector (same length as dbh), containing the
@@ -21,10 +21,10 @@
 #'   identification is available.
 #' @param new_eqtable Optional. An equation table created with the
 #'   new_equations() function.
-#' @param wna this parameter is used in the weight_allom function to determine
+#' @param wna a numeric vector, this parameter is used in the weight_allom function to determine
 #'   the dbh-related weight attributed to equations without a specified dbh
 #'   range. Default is 0.1
-#' @param w95 this parameter is used in the weight_allom function to determine
+#' @param w95 a numeric vector, this parameter is used in the weight_allom function to determine
 #'   the value at which the sample-size-related weight reaches 95% of its
 #'   maximum value (max=1). Default is 500.
 #' @param nres number of resampled values. Default is 1e4.
