+++ R/weight_allom.R
@@ -6,16 +6,16 @@
 #'
 #' @param genus a character value, containing the genus (e.g. "Quercus") of the
 #'   tree.
-#' @param coords a numerical vector of length 2 with longitude and latitude.
+#' @param coords a numeric vector of length 2 with longitude and latitude.
 #' @param species a character vector (same length as genus), containing the
 #'   species (e.g. "rubra") of the tree. Default is NULL, when no species
 #'   identification is available.
 #' @param new_eqtable Optional. An equation table created with the
 #'   new_equations() function.
-#' @param wna this parameter is used in the `weight_allom` function to determine
+#' @param wna a numeric vector, this parameter is used in the `weight_allom` function to determine
 #'   the sample-size related weights attributed to equations without a specified
 #'   sample size. Default is 0.1.
-#' @param w95 this parameter is used to determine the value at which the
+#' @param w95 a numeric vector, this parameter is used to determine the value at which the
 #'   sample-size-related weight reaches 95% of its maximum value (max=1).
 #'   Default is 500.
 #'
