+++ DESCRIPTION
@@ -19,6 +19,14 @@ Authors@R:
              family = "Anderson-Teixeira",
              role = "aut",
              email = "TeixeiraK@si.edu"))
+      person(given = "Jeffrey",
+             family = "Hanson",
+             role = "rev",
+             email = "jeffrey.hanson@uqconnect.edu.au"))
+      person(given = "Jonas",
+             family = "Stillhard",
+             role = "rev",
+             email = "jonas.stillhard@wsl.ch"))
 Description: Tool to standardize and simplify the tree biomass
     estimation process across globally distributed extratropical forests.
 License: GPL-3
