### 2021-08-19 518bb62
Revert "Added rOpenSci reviewers"
This reverts commit 60828f527e120499686fa39d692ea6f66a501084.
