+++ man/est_params.Rd
@@ -18,7 +18,7 @@ est_params(
 \item{genus}{a character vector, containing the genus (e.g. "Quercus") of
 each tree.}
 
-\item{coords}{a numerical vector of length 2 with longitude and latitude (if
+\item{coords}{a numeric vector of length 2 with longitude and latitude (if
 all trees were measured in the same location) or a matrix with 2 numerical
 columns giving the coordinates of each tree.}
 
@@ -29,18 +29,18 @@ identification is available.}
 \item{new_eqtable}{Optional. An equation table created with the
 add_equation() function. Default is the compiled allodb equation table.}
 
-\item{wna}{this parameter is used in the weight_allom() function to determine
+\item{wna}{a numeric vector, this parameter is used in the weight_allom() function to determine
 the dbh-related and sample-size related weights attributed to equations
 without a specified dbh range or sample size, respectively. Default is 0.1}
 
-\item{w95}{this parameter is used in the weight_allom() function to determine
+\item{w95}{a numeric vector, this parameter is used in the weight_allom() function to determine
 the value at which the sample-size-related weight reaches 95\% of its
 maximum value (max=1). Default is 500.}
 
 \item{nres}{number of resampled values. Default is 1e4.}
 }
 \value{
-A data frame of fitted coefficients (columns) of the non-linear
+A data frame (\code{tibble::tibble()} object) of fitted coefficients (columns) of the non-linear
 least-square regression AGB = a * dbh ^ b + e,  with e ~ N(0, sigma^2)
 }
 \description{
