+++ R/resample_agb.R
@@ -9,7 +9,7 @@
 #'   tree.
 #' @param coords a numeric vector of length 2 with longitude and latitude.
 #' @param species a character value, containing the species (e.g. "rubra") of
-#'   the tree. Default is NULL, when no species identification is available.
+#'   the tree. Default is "NULL", when no species identification is available.
 #' @param new_eqtable Optional. An equation table created with the
 #'   add_equation() function. Default is the original allodb equation table.
 #' @param wna a numeric vector, this parameter is used in the weight_allom function to determine
