### 2021-09-01 bb10a44
As suggested, using quotation marks when referring to character objects
