+++ R/illustrate_allodb.R
@@ -7,22 +7,22 @@
 #'   tree.
 #' @param coords A numeric vector of length 2 with longitude and latitude.
 #' @param species A character value, containing the species (e.g. "rubra") of
-#'   the tree. Default is `NULL`, when no species identification is available.
+#'   the tree. Default is "NULL", when no species identification is available.
 #' @param new_eqtable Optional. An equation table created with the
 #'   `add_equation()` function. Default is the base allodb equation table.
 #' @param logxy Logical: should values be plotted on a log scale? Default is
-#'   `FALSE`.
+#'   "FALSE".
 #' @param neq Number of top equations in the legend. Default is 10, meaning that
 #'   the 10 equations with the highest weights are shown in the legend.
 #' @param eqinfo Which column(s) of the equation table should be used in the
-#'   legend? Default is `"equation_taxa"`.
+#'   legend? Default is "equation_taxa".
 #' @param wna a numeric vector, this parameter is used in the weighting function to determine the
 #'   dbh-related and sample-size related weights attributed to equations without
-#'   a specified dbh range or sample size, respectively. Default is `0.1`.
+#'   a specified dbh range or sample size, respectively. Default is 0.1.
 #' @param w95 a numeric vector, this parameter is used in the weighting function to determine the
 #'   value at which the sample-size-related weight reaches 95% of its maximum
-#'   value (max=1). Default is `500`.
-#' @param nres number of resampled values. Default is `1e4`.
+#'   value (max=1). Default is 500.
+#' @param nres number of resampled values. Default is "1e4".
 #'
 #' @return A ggplot showing all resampled dbh-agb values. The top equations used
 #'   are shown in the legend. The red curve on the graph represents the final
