### 2021-09-01 2f6b69c
Added @seealso to help navigate documentation
on related functions...

Also fixed using `new_equations` function instead of `add_equation`...
