+++ R/est_params.R
@@ -14,7 +14,7 @@
 #'   species (e.g. "rubra")  of each tree. Default is "NULL", when no species
 #'   identification is available.
 #' @param new_eqtable Optional. An equation table created with the
-#'   add_equation() function. Default is the compiled allodb equation table.
+#'   new_equation() function. Default is the compiled allodb equation table.
 #' @param wna a numeric vector, this parameter is used in the weight_allom() function to determine
 #'   the dbh-related and sample-size related weights attributed to equations
 #'   without a specified dbh range or sample size, respectively. Default is 0.1
@@ -25,6 +25,9 @@
 #'
 #' @return A data frame (`tibble::tibble()` object) of fitted coefficients (columns) of the non-linear
 #'   least-square regression AGB = a * dbh ^ b + e,  with e ~ N(0, sigma^2)
+#'
+#' @seealso [weight_allom()], [new_equation()]
+#'
 #' @export
 #'
 #' @examples
