+++ R/get_biomass.R
@@ -53,6 +53,8 @@
 #' @return A vector of class "numeric" of the same length as dbh, containing AGB
 #'   value (in kg) for every stem
 #'
+#' @seealso [weight_allom())], [new_equations()]
+#'
 #' @export
 #'
 #' @examples
