### 2021-09-02 4434028
Now using \pkg{allodb} when referring to the package (as suggested)
