+++ man/resample_agb.Rd
@@ -2,7 +2,7 @@
 % Please edit documentation in R/resample_agb.R
 \name{resample_agb}
 \alias{resample_agb}
-\title{Resample allodb equations to calibrate new allometries.}
+\title{Resample \pkg{allodb} equations to calibrate new allometries.}
 \usage{
 resample_agb(
   genus,
@@ -21,10 +21,10 @@ tree.}
 \item{coords}{a numeric vector of length 2 with longitude and latitude.}
 
 \item{species}{a character value, containing the species (e.g. "rubra") of
-the tree. Default is NULL, when no species identification is available.}
+the tree. Default is "NULL", when no species identification is available.}
 
 \item{new_eqtable}{Optional. An equation table created with the
-add_equation() function. Default is the original allodb equation table.}
+new_equation() function. Default is the original \pkg{allodb}  equation table.}
 
 \item{wna}{a numeric vector, this parameter is used in the weight_allom function to determine
 the dbh-related and sample-size related weights attributed to equations
@@ -34,7 +34,7 @@ without a specified dbh range or sample size, respectively. Default is 0.1}
 the value at which the sample-size-related weight reaches 95\% of its
 maximum value (max=1). Default is 500.}
 
-\item{nres}{number of resampled values. Default is 1e4.}
+\item{nres}{number of resampled values. Default is "1e4".}
 }
 \value{
 A data frame (\code{tibble::tibble()} object) of resampled DBHs and associated AGB from the equation
@@ -42,7 +42,7 @@ table; the number of  resampled DBHs is proportional to the weight provided
 by the \code{weight_allom} function.
 }
 \description{
-After attributing a weight to each equation in allodb using the weight_allom
+After attributing a weight to each equation in \pkg{allodb}  using the weight_allom
 function, equations are then resampled within their original DBH range using
 \code{resample_agb}: the number of resampled values for each equation is
 proportional to its weight. It creates S3 objects of class "numeric".
@@ -54,4 +54,7 @@ resample_agb(
   coords = c(-78.2, 38.9)
 )
 
+}
+\seealso{
+\code{\link[=weight_allom]{weight_allom()}}, \code{\link[=new_equation]{new_equation()}}
 }
