+++ man/weight_allom.Rd
@@ -20,11 +20,11 @@ tree.}
 \item{coords}{a numeric vector of length 2 with longitude and latitude.}
 
 \item{species}{a character vector (same length as genus), containing the
-species (e.g. "rubra") of the tree. Default is NULL, when no species
+species (e.g. "rubra") of the tree. Default is "NULL", when no species
 identification is available.}
 
 \item{new_eqtable}{Optional. An equation table created with the
-new_equations() function.}
+\code{new_equations()} function.}
 
 \item{wna}{a numeric vector, this parameter is used in the \code{weight_allom} function to determine
 the sample-size related weights attributed to equations without a specified
@@ -79,4 +79,7 @@ x <- weight_allom(
 )
 str(x)
 head(x)
+}
+\seealso{
+\code{\link[=weight_allom]{weight_allom()}}, \code{\link[=new_equations]{new_equations()}}
 }
