+++ man/illustrate_allodb.Rd
@@ -2,7 +2,7 @@
 % Please edit documentation in R/illustrate_allodb.R
 \name{illustrate_allodb}
 \alias{illustrate_allodb}
-\title{Illustrate the resampling of AGB values used in allodb.}
+\title{Illustrate the resampling of AGB values used in \pkg{allodb}.}
 \usage{
 illustrate_allodb(
   genus,
@@ -24,29 +24,29 @@ tree.}
 \item{coords}{A numeric vector of length 2 with longitude and latitude.}
 
 \item{species}{A character value, containing the species (e.g. "rubra") of
-the tree. Default is \code{NULL}, when no species identification is available.}
+the tree. Default is "NULL", when no species identification is available.}
 
 \item{new_eqtable}{Optional. An equation table created with the
-\code{add_equation()} function. Default is the base allodb equation table.}
+\code{new_equation()} function. Default is the base \pkg{allodb}  equation table.}
 
 \item{logxy}{Logical: should values be plotted on a log scale? Default is
-\code{FALSE}.}
+"FALSE".}
 
 \item{neq}{Number of top equations in the legend. Default is 10, meaning that
 the 10 equations with the highest weights are shown in the legend.}
 
 \item{eqinfo}{Which column(s) of the equation table should be used in the
-legend? Default is \code{"equation_taxa"}.}
+legend? Default is "equation_taxa".}
 
 \item{wna}{a numeric vector, this parameter is used in the weighting function to determine the
 dbh-related and sample-size related weights attributed to equations without
-a specified dbh range or sample size, respectively. Default is \code{0.1}.}
+a specified dbh range or sample size, respectively. Default is 0.1.}
 
 \item{w95}{a numeric vector, this parameter is used in the weighting function to determine the
 value at which the sample-size-related weight reaches 95\% of its maximum
-value (max=1). Default is \code{500}.}
+value (max=1). Default is 500.}
 
-\item{nres}{number of resampled values. Default is \code{1e4}.}
+\item{nres}{number of resampled values. Default is "1e4".}
 }
 \value{
 A ggplot showing all resampled dbh-agb values. The top equations used
@@ -54,7 +54,7 @@ are shown in the legend. The red curve on the graph represents the final
 fitted equation.
 }
 \description{
-This function illustrates the resampling of AGB values used in allodb. It
+This function illustrates the resampling of AGB values used in \pkg{allodb}. It
 creates objects of class "ggplot".
 }
 \examples{
@@ -64,4 +64,7 @@ illustrate_allodb(
   coords = c(-78.2, 38.9)
 )
 
+}
+\seealso{
+\code{\link[=weight_allom]{weight_allom()}}, \code{\link[=new_equation]{new_equation()}}
 }

[…]
