+++ R/illustrate_allodb.R
@@ -9,17 +9,17 @@
 #' @param species A character value, containing the species (e.g. "rubra") of
 #'   the tree. Default is "NULL", when no species identification is available.
 #' @param new_eqtable Optional. An equation table created with the
-#'   `new_equation()` function. Default is the base \pkg{allodb}  equation table.
+#'   `new_equations()` function. Default is the base \pkg{allodb}  equation table.
 #' @param logxy Logical: should values be plotted on a log scale? Default is
 #'   "FALSE".
 #' @param neq Number of top equations in the legend. Default is 10, meaning that
 #'   the 10 equations with the highest weights are shown in the legend.
 #' @param eqinfo Which column(s) of the equation table should be used in the
 #'   legend? Default is "equation_taxa".
-#' @param wna a numeric vector, this parameter is used in the weighting function to determine the
+#' @param wna a numeric vector, this parameter is used in the `weight_allom()` function to determine the
 #'   dbh-related and sample-size related weights attributed to equations without
 #'   a specified dbh range or sample size, respectively. Default is 0.1.
-#' @param w95 a numeric vector, this parameter is used in the weighting function to determine the
+#' @param w95 a numeric vector, this parameter is used in the `weight_allom()` function to determine the
 #'   value at which the sample-size-related weight reaches 95% of its maximum
 #'   value (max=1). Default is 500.
 #' @param nres number of resampled values. Default is "1e4".
