+++ R/weight_allom.R
@@ -12,14 +12,14 @@
 #'   identification is available.
 #' @param new_eqtable Optional. An equation table created with the
 #'   `new_equations()` function.
-#' @param wna a numeric vector, this parameter is used in the `weight_allom` function to determine
+#' @param wna a numeric vector, this parameter is used in the `weight_allom()` function to determine
 #'   the sample-size related weights attributed to equations without a specified
 #'   sample size. Default is 0.1.
 #' @param w95 a numeric vector, this parameter is used to determine the value at which the
 #'   sample-size-related weight reaches 95% of its maximum value (max=1).
 #'   Default is 500.
 #'
-#' @details Each equation is given a weight by the function `weight_allom`,
+#' @details Each equation is given a weight by the function `weight_allom()`,
 #'   calculated as the product of the following components:
 #'
 #'   (1) sample-size weight, calculated as:
