### 2021-09-02 314dc12
Removed library(allodb) from test scripts
