+++ tests/testthat/test-get_biomass.R
@@ -1,5 +1,3 @@
-library(allodb)
-
 test_that("get_biomass works with genus and species
           information and provides reasonable outputs", {
             expect_equal(get_biomass(
