+++ tests/testthat/test-weight_allom.R
@@ -1,5 +1,3 @@
-library(allodb)
-
 test_that("weight_allom returns a vector of the same length
           as the equation table.", {
             newtab <- new_equations(subset_taxa = "Quercus")
