+++ tests/testthat/test-est_params.R
@@ -1,5 +1,3 @@
-library(allodb)
-
 test_that("est_params() returns a dataframe with all
           combinations of species x locations", {
             test_tab <- allodb::scbi_stem1[1:50, ]
