+++ man/get_biomass.Rd
@@ -31,13 +31,13 @@ species (e.g. "rubra")  of each tree. Default is "NULL", when no species
 identification is available.}
 
 \item{new_eqtable}{Optional. An equation table created with the
-new_equations() function.}
+\code{new_equations()} function.}
 
-\item{wna}{a numeric vector, this parameter is used in the weight_allom function to determine
+\item{wna}{a numeric vector, this parameter is used in the \code{weight_allom()} function to determine
 the dbh-related weight attributed to equations without a specified dbh
 range. Default is 0.1}
 
-\item{w95}{a numeric vector, this parameter is used in the weight_allom function to determine
+\item{w95}{a numeric vector, this parameter is used in the \code{weight_allom()} function to determine
 the value at which the sample-size-related weight reaches 95\% of its
 maximum value (max=1). Default is 500.}
 
@@ -63,20 +63,20 @@ each taxon (arguments \code{genus} and  \code{species}) and location (argument
 \code{coords}) in the user-provided census data. The new allometric equation is
 based on a set of allometric equations that can be customized using the
 \code{new_eqtable} argument. Each equation is then given a weight with the
-\code{weight_allom} function, based on: 1) its original sample size (numbers of
+\code{weight_allom()} function, based on: 1) its original sample size (numbers of
 trees used to develop a given allometry), 2) its climatic similarity with
 the target location, and 3) its taxonomic similarity with the target taxon
-(see documentation of the \code{weight_allom} function). The final weight
+(see documentation of the \code{weight_allom()} function). The final weight
 attributed to each equation is the product of those three weights.
-Equations are then resampled with the function \code{resample_agb}: the number
+Equations are then resampled with the\code{resample_agb()} funtion: the number
 of samples per equation is proportional to its weight, and the total number
 of samples is provided by the argument \code{nres}. The resampling is done by
 drawing DBH values from a uniform distribution on the DBH range of the
 equation, and estimating the AGB with the equation. The couples of values
-(DBH, AGB) obtained are then used in the function \code{est_params} to calibrate
+(DBH, AGB) obtained are then used in the function \code{est_params()} to calibrate
 a new allometric equation, by applying a linear regression to the
 log-transformed data. The parameters of the new allometric equations are
-then used in the \code{get_biomass} function by back-transforming the AGB
+then used in the \code{get_biomass()} function by back-transforming the AGB
 predictions based on the user-provided DBHs.
 }
 \examples{
