+++ man/illustrate_allodb.Rd
@@ -27,7 +27,7 @@ tree.}
 the tree. Default is "NULL", when no species identification is available.}
 
 \item{new_eqtable}{Optional. An equation table created with the
-\code{new_equation()} function. Default is the base \pkg{allodb}  equation table.}
+\code{new_equations()} function. Default is the base \pkg{allodb}  equation table.}
 
 \item{logxy}{Logical: should values be plotted on a log scale? Default is
 "FALSE".}
@@ -38,11 +38,11 @@ the 10 equations with the highest weights are shown in the legend.}
 \item{eqinfo}{Which column(s) of the equation table should be used in the
 legend? Default is "equation_taxa".}
 
-\item{wna}{a numeric vector, this parameter is used in the weighting function to determine the
+\item{wna}{a numeric vector, this parameter is used in the \code{weight_allom()} function to determine the
 dbh-related and sample-size related weights attributed to equations without
 a specified dbh range or sample size, respectively. Default is 0.1.}
 
-\item{w95}{a numeric vector, this parameter is used in the weighting function to determine the
+\item{w95}{a numeric vector, this parameter is used in the \code{weight_allom()} function to determine the
 value at which the sample-size-related weight reaches 95\% of its maximum
 value (max=1). Default is 500.}
 
