+++ man/est_params.Rd
@@ -27,13 +27,13 @@ species (e.g. "rubra")  of each tree. Default is "NULL", when no species
 identification is available.}
 
 \item{new_eqtable}{Optional. An equation table created with the
-new_equation() function. Default is the compiled \pkg{allodb}  equation table.}
+\code{new_equation()} function. Default is the compiled \pkg{allodb}  equation table.}
 
-\item{wna}{a numeric vector, this parameter is used in the weight_allom() function to determine
+\item{wna}{a numeric vector, this parameter is used in the \code{weight_allom()} function to determine
 the dbh-related and sample-size related weights attributed to equations
 without a specified dbh range or sample size, respectively. Default is 0.1}
 
-\item{w95}{a numeric vector, this parameter is used in the weight_allom() function to determine
+\item{w95}{a numeric vector, this parameter is used in the \code{weight_allom()} function to determine
 the value at which the sample-size-related weight reaches 95\% of its
 maximum value (max=1). Default is 500.}
 
