### 2021-09-02 eeea216
Fixed minor mistakes
