+++ R/resample_agb.R
@@ -11,7 +11,7 @@
 #' @param species a character value, containing the species (e.g. "rubra") of
 #'   the tree. Default is "NULL", when no species identification is available.
 #' @param new_eqtable Optional. An equation table created with the
-#'   `new_equation()` function. Default is the original \pkg{allodb}  equation table.
+#'   `new_equations()` function. Default is the original \pkg{allodb}  equation table.
 #' @param wna a numeric vector, this parameter is used in the `weight_allom()` function to determine
 #'   the dbh-related and sample-size related weights attributed to equations
 #'   without a specified dbh range or sample size, respectively. Default is 0.1
@@ -24,7 +24,7 @@
 #'   table; the number of  resampled DBHs is proportional to the weight provided
 #'   by the `weight_allom()` function.
 #'
-#' @seealso [weight_allom()], [new_equation()]
+#' @seealso [weight_allom()], [new_equations()]
 #'
 #' @export
 #'
