+++ man/weight_allom.Rd
@@ -81,5 +81,5 @@ str(x)
 head(x)
 }
 \seealso{
-\code{\link[=weight_allom]{weight_allom()}}, \code{\link[=new_equations]{new_equations()}}
+\code{\link[=get_biomass]{get_biomass()}}, \code{\link[=new_equations]{new_equations()}}
 }
