+++ man/resample_agb.Rd
@@ -24,7 +24,7 @@ tree.}
 the tree. Default is "NULL", when no species identification is available.}
 
 \item{new_eqtable}{Optional. An equation table created with the
-\code{new_equation()} function. Default is the original \pkg{allodb}  equation table.}
+\code{new_equations()} function. Default is the original \pkg{allodb}  equation table.}
 
 \item{wna}{a numeric vector, this parameter is used in the \code{weight_allom()} function to determine
 the dbh-related and sample-size related weights attributed to equations
@@ -56,5 +56,5 @@ resample_agb(
 
 }
 \seealso{
-\code{\link[=weight_allom]{weight_allom()}}, \code{\link[=new_equation]{new_equation()}}
+\code{\link[=weight_allom]{weight_allom()}}, \code{\link[=new_equations]{new_equations()}}
 }
