+++ R/est_params.R
@@ -24,7 +24,7 @@
 #' @param nres number of resampled values. Default is "1e4".
 #'
 #' @return A data frame (`tibble::tibble()` object) of fitted coefficients (columns) of the non-linear
-#'   least-square regression AGB = a * dbh ^ b + e,  with e ~ N(0, sigma^2)
+#'   least-square regression \deqn{AGB = a * dbh ^ b + e,  with e ~ N(0, sigma^2)}
 #'
 #' @seealso [weight_allom()], [new_equations()]
 #'
