+++ R/weight_allom.R
@@ -24,7 +24,7 @@
 #'
 #'   (1) sample-size weight, calculated as:
 #'
-#'   1-exp(-n*(log(20)/w95))
+#'   \deqn{1-exp(-n*(log(20)/w95))}
 #'
 #'   where n is the sample size of the equation; the weight given to equations
 #'   with no sample size information is determined by argument `wna` (0.1 by
