+++ man/shrub_species.Rd
@@ -15,9 +15,9 @@ Genus and species of shrubby plants identified in the 24 extratropical
 ForestGEO sites used in allodb.
 }
 \examples{
-head(shrub_species)
+#preview the dataset
+print(head(shrub_species))
 
-str(shrub_species)
 }
 \seealso{
 Other datasets: 
