+++ man/koppenMatrix.Rd
@@ -30,7 +30,9 @@ zone). The resulting weight has a value between 0 (different climate groups)
 and 1 (exactly the same climate classification).
 }
 \examples{
-koppenMatrix
+#preview the dataset
+print(head(koppenMatrix))
+
 }
 \seealso{
 Other datasets: 
