+++ man/sitespecies.Rd
@@ -23,13 +23,14 @@ their tree species.
 }
 }
 \examples{
-sitespecies
+#preview the datasets
+print(head(sitespecies))
+print(head(sitespecies_metadata))
 
-sitespecies_metadata
 }
 \seealso{
 Other database datasets: 
-\code{\link{equations}},
+\code{\link{equations_metadata}},
 \code{\link{missing_values}},
 \code{\link{references}},
 \code{\link{sites_info}}
