+++ man/equations.Rd
@@ -1,21 +1,16 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/data.R
 \docType{data}
-\name{equations}
-\alias{equations}
+\name{equations_metadata}
 \alias{equations_metadata}
 \title{Tables of allometric equations and associated metadata.}
 \format{
-An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 570 rows and 47 columns.
-
 An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 47 rows and 7 columns.
 }
 \source{
 See \link{references} for equations original sources.
 }
 \usage{
-equations
-
 equations_metadata
 }
 \description{
@@ -30,9 +25,9 @@ above-ground biomass (AGB) per species based on extratropical ForestGEO
 sites.
 }
 \examples{
-equations
-
-equations_metadata
+#preview the datasets
+print(head(equations))
+print(head(equations_metadata))
 }
 \seealso{
 Other database datasets: 
