### 2021-09-04 #436 @gonzalezeb
Hi @adamhsparks, I wasn’t sure how to respond to the multiples reviews so I hope the following is sufficient. We are very grateful for the thoughtful suggestions by reviewers, and have implemented most.

Initial suggestions by [@jeffreyhanson (5/19/21)](https://github.com/ropensci/software-review/issues/436#issuecomment-844578387), have been addressed:
1.	The package website is working (https://forestgeo.github.io/allodb/)
2.	Documentation for multiple functions has been improved by using code formatting (eg. using ‘numeric vector’ instead of ‘numerical vector’, added (`tibble::tibble()` object), etc).
3.	Typos have been corrected
4.	Unnecessary R and non-R script files have been removed (eg. R/sysdata.rda)
5.	We agree with the reviewer. If a user passes a bad argument, it is best to fail fast with an informative message so that the user can quickly correct their mistake and move on. But the number of possible mistakes may be infinite; so we have particularly focused on adding an internal function `validate_equations()`to tests that all equations in `equations$equation_allometry` are a valid R expression of the dbh or h variables (since this could be one of the main mistakes on calculating biomass). This function can be called from inside any function within `allodb` and allows users to pass a character vector of equations. We also added consistency checks in the get_biomass function of dbh and coordinates provided by the user.
6.	Unused Travis badge in REAMDE has been removed
7.	The .buildignore folder has been removed
8.	We changed `set.seed` with `withr::with_seed` in the `resample_agb` function.
9.	About `eval(parse(…))`, see next comment.
10.	Parameters for many functions have been reordered.

Comprehensive review by [@jeffreyhanson (6/17/21)](https://github.com/ropensci/software-review/issues/436#issuecomment-863704236), has also been addressed:
**About code:**
1.	See note number 5 above. If the suggested change is truly necessary (validation of function arguments), then we may take the time to implement.
2.	All dataframes now have the `tibble` subclass.
**About documentation:**
3.	Documentation has been improved for all functions (specially the `get_biomass` and `weight_allom` functions), added `@seealso` to help users navigate btw documentations, 	use quotation marks for character vectors, added example sections to the documentation for the built-in datasets, etc.
4.	We are now using \pkg{allodb} in documentation when referring to this package, and equations commands (eg. \deqn{}).
5.	Removed library(allodb) from all test scripts.

Suggestions by [@jstillh (6/14/21)](https://github.com/ropensci/software-review/issues/436#issuecomment-860737938), have been addressed:
-	Documentation for both the `get_biomass` and `weight_allom` core functions have been improved so the weighting concept is more transparent, specifically we give more details in the Details section. Regarding adding additional outputs to the `get_biomass` function, we think that would make the function too slow.
-	We improved the error message within the `weight_allom` function (i.e. now it deals with coordinates that are not associated with any koppen climate zone) to address the specific comments on the fact that the get_biomass function throws an error when the equation table does not contain equations corresponding to the Koppen climate at the provided coordinates (argument coords). The proposed solution is to change koppenMatrix, replacing climate combinations with weight we=0 (2 completely different climates) to we = 1e-6, similar to what is done with the taxonomic weight. This way, equations with we = 1e-6 will only be significantly used in the recalibrated equations when no better climate combination is found, therefore avoid getting an error. In regards to Forrester et al 2017 equations, we revised them and now many coordinates correspond to max or min values given in original publication (not the mean except for one site).
-	Vignette has been improved.
-  About eval(parse(…)), also see next comment.
