### 2021-09-05 db045af
Update allodb-vignette.Rmd
