+++ vignettes/allodb-vignette.Rmd
@@ -155,7 +155,7 @@ plot(
 
 ### Calibrate a new equation for each taxon/location combination
 
-The resampled values are then used to fit the following nonlinear model: $AGB = a \cdot dbh ^ b + e$, with i.i.d. $e \sim \mathcal{N}(0, sigma^2)$. The parameters (_a_, _b_, and _sigma_) are returned by the `est_params()` function. 
+The resampled values are then used to fit the following nonlinear model: $AGB = a \cdot dbh ^ b + e$, with i.i.d. $e \sim \mathcal{N}(0, sigma^2)$. The parameters (_a_, _b_, and _sigma_) are returned by the `est_params()` function. In other words, this function calibrates new allometric equations from sampling previous ones. New allometric equations are calibrated for each species and location by resampling the original compiled equations; equations with a larger sample size, and/or higher taxonomic rank, and climatic similarity with the species and location in question are given a higher weight in this process.
 
 ```{r est-params-acer, fig.width = 6, fig.height = 5}
 pars_acer <- est_params(
@@ -188,7 +188,7 @@ AGB is then recalculated as `AGB = a * dbh^b` within the `get_biomass()` functio
 
 ## Visualize the recalibration of equations
 
-The recalibrated equation for one taxon/location combination can be easily visualized with the `illustrate_allodb()` function, which returns a ggplot (see package `ggplot2`) with all resampled values, and the top equations used in the legend. The user can control the number of equations and equation information displayed in the legend. The red dotted line is the recalibrated equation used in the `get_biomass()` function. 
+The recalibrated equation for one taxon/location combination can be easily visualized with the `illustrate_allodb()` function, which returns a ggplot (see package `ggplot2`) with all resampled values, and the top equations used displayed in the legend. The user can control the number of equations and equation information shown in the legend. The red dotted line is the recalibrated equation used in the `get_biomass()` function. 
 
 ```{r illustrate-allodb, fig.height = 4, fig.width=8}
 illustrate_allodb(
