### 2021-09-05 1cf8d2e
Update scbi_stem1.Rd
