### 2021-09-08 #436 @adamhsparks
@gonzalezeb, thank you for your comprehensive overview of the changes that you have made to the _allodb_ package.

@jeffreyhanson and @jstillh could you please let me know if you are satisfied with the changes to the package or have any further suggestions that may improve it?
