### 2021-09-14 #436 @jeffreyhanson
Thank you @gonzalezeb for providing such detailed responses to the points I raised earlier. I especially found your discussion on the costs and benefits of `eval(parse(...))` really useful. I think the R package is looking absolutely brilliant! Most of the points I raised earlier have been addressed. I've made a couple of suggestions below that I think could help improve package.<sup>1</sup> All things considered, these are pretty minor - just some documentation tweaks. If those could be addressed, I'd say that all my suggestions have been satisfied. I've also listed several additional minor comments too. Based on my understanding of rOpenSci guidelines, they won't/shouldn't influence whether the package should be accepted or not (but please correct me if I am wrong). I've listed them just to be thorough, and in case the package authors think they are worth addressing.

### Suggestions

* Package checks throw the following warning on my computer (i.e. running `devtools::check()`). Could you please fix it? Let me know if you're unable to reproduce this error and I can try to provide more information?

    >  Undocumented code objects:
    >    ‘equations’
    >  Undocumented data sets:
    >    ‘equations’
    >  All user-level objects in a package should have documentation entries.
    >  See chapter ‘Writing R documentation files’ in the ‘Writing R
    >  Extensions’ manual.

* Thank you very much for looking into potential issues with using `eval(parse(...))` and potential alternatives. I agree that the main issue would be if the code was run inside a web application that allowed individuals to enter in arbitrary text. Although additional issues could arise on standard desktop/laptop computers too (e.g. a malicious actor could potentially update the library files), I guess at this point malicious actors would have easier ways to accomplish their goals then modifying R library files. So, given the costs and benefits outlined by @gonzalezeb, I think it's perfectly reasonable for the package to use `eval(paste(...))`. Maybe it would be worth adding some warning text to the documentation for the function(s) that use `eval(parse(...))` so that users are aware of potential issues? This way it is clearly the responsibility of the user to ensure that they take appropriate steps to avoid security issues.<sup>2</sup>

    Also, I noticed one of the links provided by @gonzalezeb mentioned that [`RAppArmor::eval.secure()`](https://www.rdocumentation.org/packages/RAppArmor/versions/2.0.2/topics/eval.secure) can prevent system calls with administrator privileges. If it's not too much hassle, would it be worth using that where possible? It seems that this R package is only available on Linux though - so perhaps it's not worth the hassle (because then some additional code would be needed to conditionally use `RAppArmor::eval.secure()` if it's available, and otherwise use `eval(parse(...))`). I just thought I'd mention it just in case.

* The README.md file might need re-knitting because the output displayed for `est_params()` is a `data.table` object (instead of a `tibble()` object per the current version).

### Minor comments

* The documentation doesn't seem to format code correctly (i.e. using back ticks). For example, NULL is referred to as "NULL" instead of `NULL` (e.g., see [here](https://forestgeo.github.io/allodb/reference/get_biomass.html), or [here](https://forestgeo.github.io/allodb/reference/resample_agb.html)). Personally, I find syntax formatting helpful when reading documentation (e.g., https://readr.tidyverse.org/reference/read_delim.html). I don't think rOpensci guidelines have style mandates for code in package documentation, so I don't think this is critically important.

* The main functions (e.g.`est_params()`, `get_biomass()`, `resample_agb()`) don't contain code to validate all arguments. Although @gonzalezeb mention that "the number of possible mistakes may be infinite", I would suggest that perhaps the number of expectations (e.g. properties or characteristics) for a valid input is much lower. For example, in order to verify that an argument is a single integer (e.g. `5`), one does not need to verify that the parameter is (i) not a character, (ii) not a `NULL`, (iii) not a double, (Inf) etc. Instead, one could verify that it conforms to a set of expectations (e.g. `inherits(x, "integer") && identical(length(x), 1) && all(is.finite(x))`). This is made much easier using existing R packages developed for this purpose though. However, I don't think rOpensci or CRAN require argument validation for published packages, so I don't think this issue critically important. I'm just making this comment to be thorough :)

* This is a super minor suggestion. The equation in the documentation for the `est_params()` function has the text " with " formatted as a maths term (rather than text), which makes it slightly difficult to read on the [package website]( https://forestgeo.github.io/allodb/reference/est_params.html). To address this, I think you could change [this line](https://github.com/forestgeo/allodb/blob/master/R/est_params.R#L27) to define the equation as:
```
\deqn{AGB = a * dbh ^ b + e, \space \mathit{with} \space e ~ N(0, sigma^2)}{AGB = a * dbh ^ b + e, with e ~ N(0, sigma^2)}
```

-------------------

<sup>1</sup> Please note that I still (unsurprisingly) lack the forestry experience needed to evaluate the novelty and methodology of this package, and so continue to limit my reviews to code quality and documentation.

<sup>2</sup> Just to be clear, I do think that -- where possible, feasible, and reasonable -- R package developers should ideally -- if only in principle -- minimize security risks, even if software licenses absolve developers of responsibility. It's just I think that `eval(paste(..))` is not so terrible here, given the noted trade-offs that are specific to this package. Perhaps other will disagree, and that's fine. Please note that I do not have any experience with cybersecurity. If you want an informed view point on such matters, I suggest you talk to someone else :)
