### 2021-09-14 b464850
Updated data.R
