+++ R/data.R
@@ -88,7 +88,7 @@
 #'
 #' A table with tree data from the Smithsonian Conservation Biology Institute,
 #' USA (SCBI) ForestGEO dynamics plot. This dataset is an extract from the first tree census
-#' in 2008, only covering 1 hectare (SCBI is 25.6 ha). DBH in cm.
+#' in 2008, only covering 1 hectare (SCBI plot is 25.6 ha). DBH in cm.
 #'
 #' @source Full datasets for tree census data at SCBI can be requested through
 #'   the ForestGEO portal (<https://forestgeo.si.edu/>). Census 1, 2, and 3 can
