### 2021-09-15 0265f9b
Added an example to get biomass at a site with multiple coordinates
this point was brought out by a manuscript reviewer
