+++ R/get_biomass.R
@@ -58,10 +58,16 @@
 #' @export
 #'
 #' @examples
-#' # Estimate the biomass of all individuals from the Lauraceae family
+#' # Estimate the biomass of all individuals from the Lauraceae family at the SCBI plot
 #' lau <- subset(scbi_stem1, Family == "Lauraceae")
-#' agb <- get_biomass(lau$dbh, lau$genus, lau$species, coords = c(-78.2, 38.9))
-#' str(agb)
+#' lau$agb <- get_biomass(lau$dbh, lau$genus, lau$species, coords = c(-78.2, 38.9))
+#'
+#' # Estimate biomass simultaneously from multiple sites (here using scbi_stem1 data as example with multiple coordinates)
+#' dat = scbi_stem1[1:100, ]
+#' dat$long=c(rep(-78,50),rep(-80,50))
+#' dat$lat = c(rep(40, 50), rep(41, 50))
+#' dat$biomass<-get_biomass( dbh = dat$dbh, genus=dat$genus, species = dat$species, coords = dat[, c("long", "lat")])
+#'
 get_biomass <- function(dbh,
                         genus,
                         coords,
