+++ man/scbi_stem1.Rd
@@ -19,7 +19,7 @@ scbi_stem1
 \description{
 A table with tree data from the Smithsonian Conservation Biology Institute,
 USA (SCBI) ForestGEO dynamics plot. This dataset is an extract from the first tree census
-in 2008, only covering 1 hectare (SCBI is 25.6 ha). DBH in cm.
+in 2008, only covering 1 hectare (SCBI plot is 25.6 ha). DBH in cm.
 }
 \examples{
 #preview the datasets
