### 2021-09-15 5b4d091
Fixed documentation for equation table, now working
