+++ man/sites_info.Rd
@@ -25,7 +25,7 @@ print(head(sites_info))
 }
 \seealso{
 Other database datasets: 
-\code{\link{equations_metadata}},
+\code{\link{equations}},
 \code{\link{missing_values}},
 \code{\link{references}},
 \code{\link{sitespecies}}
