+++ man/references.Rd
@@ -33,7 +33,7 @@ print(head(references_metadata))
 }
 \seealso{
 Other database datasets: 
-\code{\link{equations_metadata}},
+\code{\link{equations}},
 \code{\link{missing_values}},
 \code{\link{sites_info}},
 \code{\link{sitespecies}}
