### 2021-09-16 98c9b55
Updating readme so a tibble is shown now for est_params
Deleted ![](docs/figures/resample-acer-1.png) because the file doesn't exist..
