+++ README.Rmd
@@ -121,7 +121,6 @@ plot(
 )
 ```
 
-![](docs/figures/resample-acer-1.png)
 
 The resampled values are then used to fit the following nonlinear model: <img src="https://render.githubusercontent.com/render/math?math=AGB = a * dbh ^ b %2B e">, with i.i.d. <img src="https://render.githubusercontent.com/render/math?math=e ~N(0, sigma^2)">. The parameters (_a_, _b_, and _sigma_) are returned by the `est_params()` function.
 
@@ -140,7 +139,6 @@ illustrate_allodb(
 )
 ```
 
-![](docs/figures/est-params-acer-1.png)
 
 The `est_params` function can be used for all species/site combinations in the dataset at once. 
 
