+++ README.md
@@ -1,22 +1,54 @@
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
+
 # <img src="https://i.imgur.com/39pvr4n.png" align="left" height=44 /> allodb: An R package for biomass estimation at extratropical forest plots
 
 <!-- badges: start -->
-[![lifecycle](https://img.shields.io/badge/lifecycle-experimental-orange.svg)](https://lifecycle.r-lib.org/articles/stages.html)
-[![Coverage status](https://coveralls.io/repos/github/forestgeo/allodb/badge.svg)](https://coveralls.io/github/forestgeo/allodb)
-[![R-CMD-check](https://github.com/forestgeo/allodb/workflows/R-CMD-check/badge.svg)](https://github.com/forestgeo/allodb/actions)
+
+[![lifecycle](https://img.shields.io/badge/lifecycle-experimental-orange.svg)](https://lifecycle.r-lib.org/articles/stages.html)  
+[![Build
+Status](https://travis-ci.org/forestgeo/allodb.svg?branch=master)](https://travis-ci.org/forestgeo/allodb)  
+[![Coverage
+status](https://coveralls.io/repos/github/forestgeo/allodb/badge.svg)](https://coveralls.io/github/forestgeo/allodb)  
+[![R-CMD-check](https://github.com/forestgeo/allodb/workflows/R-CMD-check/badge.svg)](https://github.com/forestgeo/allodb/actions)  
+[![peer-review](https://badges.ropensci.org/436_status.svg)](https://github.com/ropensci/software-review/issues/436)
 <!-- badges: end -->
 
 ## Introduction
 
-Allometric equations for calculation of tree aboveground biomass (AGB) form the basis for estimates of forest carbon storage and exchange with the atmosphere. While standard models exist to calculate forest biomass across the tropics, we lack a standardized tool for computing AGB across the global extratropics.
+Allometric equations for calculation of tree aboveground biomass (AGB)
+form the basis for estimates of forest carbon storage and exchange with
+the atmosphere. While standard models exist to calculate forest biomass
+across the tropics, we lack a standardized tool for computing AGB across
+the global extratropics.
 
-*allodb* was conceived as a framework to standardize and simplify the biomass estimation process across globally distributed extratropical forests (mainly temperate and boreal forests). With *allodb* we aimed to: a) compile relevant published and unpublished allometries, focusing on AGB but structured to handle other variables (e.g., height); b) objectively select and integrate appropriate available equations across the full range of tree sizes; and c) serve as a platform for future updates and expansion to other research sites.
+*allodb* was conceived as a framework to standardize and simplify the
+biomass estimation process across globally distributed extratropical
+forests (mainly temperate and boreal forests). With *allodb* we aimed
+to: a) compile relevant published and unpublished allometries, focusing
+on AGB but structured to handle other variables (e.g., height); b)
+objectively select and integrate appropriate available equations across
+the full range of tree sizes; and c) serve as a platform for future
+updates and expansion to other research sites.
 
[…]
