+++ R/new_equations.R
@@ -44,7 +44,7 @@
 #'   is "Total aboveground biomass".
 #' @param use_height_allom a logical value. In \pkg{allodb} we use Bohn et al. (2014)
 #'   for European sites. User need to provide height allometry when needed to
-#'   calculate AGB. Default is "TRUE".
+#'   calculate AGB. Default is `TRUE`.
 #'
 #' @return A new equation dataframe (`tibble::tibble()` object).
 #'
