+++ R/illustrate_allodb.R
@@ -7,11 +7,11 @@
 #'   tree.
 #' @param coords A numeric vector of length 2 with longitude and latitude.
 #' @param species A character value, containing the species (e.g. "rubra") of
-#'   the tree. Default is "NULL", when no species identification is available.
+#'   the tree. Default is `NULL`, when no species identification is available.
 #' @param new_eqtable Optional. An equation table created with the
 #'   `new_equations()` function. Default is the base \pkg{allodb}  equation table.
 #' @param logxy Logical: should values be plotted on a log scale? Default is
-#'   "FALSE".
+#'   `FALSE`.
 #' @param neq Number of top equations in the legend. Default is 10, meaning that
 #'   the 10 equations with the highest weights are shown in the legend.
 #' @param eqinfo Which column(s) of the equation table should be used in the
