### 2021-09-16 d17f7b3
Changed equation line as suggested so it can read better in package website
