### 2021-09-16 cc67cfa
Update new_equations.R
