+++ man/get_biomass.Rd
@@ -80,11 +80,11 @@ then used in the \code{get_biomass()} function by back-transforming the AGB
 predictions based on the user-provided DBHs.
 }
 \examples{
-# Estimate the biomass of all individuals from the Lauraceae family at the SCBI plot
+# Estimate biomass of all individuals from the Lauraceae family at the SCBI plot
 lau <- subset(scbi_stem1, Family == "Lauraceae")
 lau$agb <- get_biomass(lau$dbh, lau$genus, lau$species, coords = c(-78.2, 38.9))
 
-# Estimate biomass simultaneously from multiple sites (here using scbi_stem1 data as example with multiple coordinates)
+# Estimate biomass from multiple sites (using scbi_stem1 as example with multiple coord)
 dat = scbi_stem1[1:100, ]
 dat$long=c(rep(-78,50),rep(-80,50))
 dat$lat = c(rep(40, 50), rep(41, 50))
