### 2021-09-16 #436 @gonzalezeb
Thanks @jeffreyhanson, I have adopted most of your comments, they are very helpful!
