### 2021-09-16 forestgeo/allodb#147 @gonzalezeb
issue solved
