### 2021-09-17 de75b11
Reduced code lines
