### 2021-09-17 3365a7d
Reduced code lines after running gp and lint
