+++ R/get_biomass.R
@@ -10,63 +10,70 @@
 #' (usually several hundred thousand observations).
 #'
 #' @param dbh a numeric vector containing tree diameter at breast height (dbh)
-#'   measurements, in cm.
+#' measurements, in cm.
 #' @param genus a character vector (same length as dbh), containing the genus
-#'   (e.g. "Quercus") of each tree.
+#' (e.g. "Quercus") of each tree.
 #' @param coords a numeric vector of length 2 with longitude and latitude (if
-#'   all trees were measured in the same location) or a matrix with 2 numerical
-#'   columns giving the coordinates of each tree.
+#' all trees were measured in the same location) or a matrix with 2 numerical
+#' columns giving the coordinates of each tree.
 #' @param species a character vector (same length as dbh), containing the
-#'   species (e.g. "rubra")  of each tree. Default is `NULL`, when no species
-#'   identification is available.
+#' species (e.g. "rubra")  of each tree. Default is `NULL`, when no species
+#' identification is available.
 #' @param new_eqtable Optional. An equation table created with the
-#'   `new_equations()` function.
-#' @param wna a numeric vector, this parameter is used in the `weight_allom()` function to determine
-#'   the dbh-related weight attributed to equations without a specified dbh
-#'   range. Default is 0.1
-#' @param w95 a numeric vector, this parameter is used in the `weight_allom()` function to determine
-#'   the value at which the sample-size-related weight reaches 95% of its
-#'   maximum value (max=1). Default is 500.
+#' `new_equations()` function.
+#' @param wna a numeric vector, this parameter is used in the `weight_allom()`
+#' function to determine
+#' the dbh-related weight attributed to equations without a specified dbh
+#' range. Default is 0.1
+#' @param w95 a numeric vector, this parameter
+#' is used in the `weight_allom()` function to determine
+#' the value at which the sample-size-related weight reaches 95% of its
+#' maximum value (max=1). Default is 500.
 #' @param nres number of resampled values. Default is "1e4".
 #'
 #' @details `allodb` estimates AGB by calibrating a new allometric equation for
-#'   each taxon (arguments `genus` and  `species`) and location (argument
-#'   `coords`) in the user-provided census data. The new allometric equation is
-#'   based on a set of allometric equations that can be customized using the
-#'   `new_eqtable` argument. Each equation is then given a weight with the
-#'   `weight_allom()` function, based on: 1) its original sample size (numbers of
-#'   trees used to develop a given allometry), 2) its climatic similarity with
-#'   the target location, and 3) its taxonomic similarity with the target taxon
-#'   (see documentation of the `weight_allom()` function). The final weight
-#'   attributed to each equation is the product of those three weights.
-#'   Equations are then resampled with the`resample_agb()` funtion: the number
-#'   of samples per equation is proportional to its weight, and the total number
-#'   of samples is provided by the argument `nres`. The resampling is done by
-#'   drawing DBH values from a uniform distribution on the DBH range of the
-#'   equation, and estimating the AGB with the equation. The couples of values
-#'   (DBH, AGB) obtained are then used in the function `est_params()` to calibrate
-#'   a new allometric equation, by applying a linear regression to the
-#'   log-transformed data. The parameters of the new allometric equations are
-#'   then used in the `get_biomass()` function by back-transforming the AGB
-#'   predictions based on the user-provided DBHs.
+#' each taxon (arguments `genus` and  `species`) and location (argument
[…]
