+++ R/validate_equation.R
@@ -52,4 +52,3 @@ walk_apply <- function(.x, .f, ...) {
   lapply(.x, .f, ...)
   invisible(.x)
 }
-
