+++ R/new_equations.R
@@ -42,9 +42,10 @@
 #'   is "DBH", other option is "DBH, H".
 #' @param new_output_var dependent variable estimated by the allometry. Default
 #'   is "Total aboveground biomass".
-#' @param use_height_allom a logical value. In \pkg{allodb} we use Bohn et al. (2014)
-#'   for European sites. User need to provide height allometry when needed to
-#'   calculate AGB. Default is `TRUE`.
+#' @param use_height_allom a logical value. In \pkg{allodb}
+#' we use Bohn et al. (2014)
+#' for European sites. User need to provide height allometry
+#' when needed to calculate AGB. Default is `TRUE`.
 #'
 #' @return A new equation dataframe (`tibble::tibble()` object).
 #'
