+++ man/new_equations.Rd
@@ -76,9 +76,10 @@ is "DBH", other option is "DBH, H".}
 \item{new_output_var}{dependent variable estimated by the allometry. Default
 is "Total aboveground biomass".}
 
-\item{use_height_allom}{a logical value. In \pkg{allodb} we use Bohn et al. (2014)
-for European sites. User need to provide height allometry when needed to
-calculate AGB. Default is \code{TRUE}.}
+\item{use_height_allom}{a logical value. In \pkg{allodb}
+we use Bohn et al. (2014)
+for European sites. User need to provide height allometry
+when needed to calculate AGB. Default is \code{TRUE}.}
 }
 \value{
 A new equation dataframe (\code{tibble::tibble()} object).
