### 2021-09-19 #436 @jstillh
Thank you @gonzalezeb for the very detailed response to the reviews. As @jeffreyhanson, I like your detailed response regarding the use of `eval(parse(...))`. In general, i am very pleased with the revisions and the changes you implemented in the functions. I do not have any more suggestions beyond the ones by @jeffreyhanson.
- I appreciate the changes / improvements in the documentation of both the `weight_allom` and the `get_biomass` function and that you implemented an informative error message if coordinates are not covered by the koppen data.
- Your proposition regarding the weighting of the climatic combinations sounds like a good idea.

Thanks for this contribution.
