### 2021-09-27 #436 @ropensci-review-bot
Approved! Thanks @gonzalezeb for submitting and @jeffreyhanson, @jstillh for your reviews! :grin:
